Categorical algebra

Catlab.CategoricalAlgebra.FreeDiagrams — Module

Free diagrams in a category.

A free diagram in a category is a diagram whose shape is a free category. Examples include the empty diagram, pairs of objects, discrete diagrams, parallel morphisms, spans, and cospans. Limits and colimits are most commonly taken over free diagrams.

Catlab.CategoricalAlgebra.FreeDiagrams.Cospan — Type

Cospan of morphisms in a category.

A common special case of Multicospan. See also Span.

Catlab.CategoricalAlgebra.FreeDiagrams.DiscreteDiagram — Type

Discrete diagram: a diagram whose only morphisms are identities.

Catlab.CategoricalAlgebra.FreeDiagrams.FixedShapeFreeDiagram — Type

Abstract type for free diagram of fixed shape.

Catlab.CategoricalAlgebra.FreeDiagrams.Multicospan — Type

Multicospan of morphisms in a category.

A multicospan is like a Cospan except that it may have a number of legs different than two. A limit of this shape is a pullback.

Catlab.CategoricalAlgebra.FreeDiagrams.Multispan — Type

Multispan of morphisms in a category.

A multispan is like a Span except that it may have a number of legs different than two. A colimit of this shape is a pushout.

Catlab.CategoricalAlgebra.FreeDiagrams.ParallelMorphisms — Type

Parallel morphims in a category.

Parallel morphisms are just morphisms with the same domain and codomain. A (co)limit of this shape is a (co)equalizer.

For the common special case of two morphisms, see ParallelPair.

Catlab.CategoricalAlgebra.FreeDiagrams.ParallelPair — Type

Pair of parallel morphisms in a category.

A common special case of ParallelMorphisms.

Catlab.CategoricalAlgebra.FreeDiagrams.Span — Type

Span of morphims in a category.

A common special case of Multispan. See also Cospan.

Catlab.CategoricalAlgebra.Limits — Module

Limits and colimits in a category.

Catlab.CategoricalAlgebra.Limits.AbstractColimit — Type

Abstract type for colimit in a category.

The standard concrete subtype is Colimit, although for computational reasons certain categories may use different subtypes to include extra data.

Catlab.CategoricalAlgebra.Limits.AbstractLimit — Type

Abstract type for limit in a category.

The standard concrete subtype is Limit, although for computational reasons certain categories may use different subtypes to include extra data.

Catlab.CategoricalAlgebra.Limits.Colimit — Type

Colimit in a category.

Catlab.CategoricalAlgebra.Limits.Limit — Type

Limit in a category.

Catlab.CategoricalAlgebra.Limits.colimit — Function

Colimit of a diagram.

To define colimits in a category with objects Ob, override the method colimit(::FreeDiagram{Ob}) for general colimits or colimit(::D) with suitable type D <: FixedShapeFreeDiagram{Ob} for colimits of specific shape, such as coproducts or coequalizers.

See also: limit

Catlab.CategoricalAlgebra.Limits.composite_pullback — Method

Compute pullback as composite of product and equalizer.

Catlab.CategoricalAlgebra.Limits.composite_pushout — Method

Compute pushout as composite of coproduct and coequalizer.

Catlab.CategoricalAlgebra.Limits.limit — Function

Limit of a diagram.

To define limits in a category with objects Ob, override the method limit(::FreeDiagram{Ob}) for general limits or limit(::D) with suitable type D <: FixedShapeFreeDiagram{Ob} for limits of specific shape, such as products or equalizers.

See also: colimit

Catlab.CategoricalAlgebra.Limits.pullback — Method

Pullback of a pair of morphisms with common codomain.

To implement for a type T, define the method limit(::Cospan{T}) and/or limit(::Multicospan{T}) or, if you have already implemented products and equalizers, rely on the default implementation.

Catlab.CategoricalAlgebra.Limits.pushout — Method

Pushout of a pair of morphisms with common domain.

To implement for a type T, define the method colimit(::Span{T}) and/or colimit(::Multispan{T}) or, if you have already implemented coproducts and coequalizers, rely on the default implementation.

Catlab.CategoricalAlgebra.Limits.universal — Function

Universal property of (co)limits.

Compute the morphism whose existence and uniqueness is guaranteed by the universal property of (co)limits.

See also: limit, colimit.

Catlab.Theories.copair — Method

Copairing of morphisms: universal property of coproducts/pushouts.

To implement for coproducts of type T, define the method universal(::BinaryCoproduct{T}, ::Cospan{T}) and/or universal(::Coproduct{T}, ::Multicospan{T}) and similarly for pushouts.

Catlab.Theories.coproduct — Method

Coproduct of objects.

To implement for a type T, define the method colimit(::ObjectPair{T}) and/or colimit(::DiscreteDiagram{T}).

Catlab.Theories.create — Method

Unique morphism out of an initial object.

To implement for a type T, define the method universal(::Initial{T}, ::SMulticospan{0,T}).

Catlab.Theories.delete — Method

Unique morphism into a terminal object.

To implement for a type T, define the method universal(::Terminal{T}, ::SMultispan{0,T}).

Catlab.Theories.factorize — Method

Factor morphism through (co)equalizer, via the universal property.

To implement for equalizers of type T, define the method universal(::Equalizer{T}, ::SMultispan{1,T}). For coequalizers of type T, define the method universal(::Coequalizer{T}, ::SMulticospan{1,T}).

Catlab.Theories.initial — Method

Initial object.

To implement for a type T, define the method colimit(::EmptyDiagram{T}).

Catlab.Theories.pair — Method

Pairing of morphisms: universal property of products/pullbacks.

To implement for products of type T, define the method universal(::BinaryProduct{T}, ::Span{T}) and/or universal(::Product{T}, ::Multispan{T}) and similarly for pullbacks.

Catlab.Theories.product — Method

Product of objects.

To implement for a type T, define the method limit(::ObjectPair{T}) and/or limit(::DiscreteDiagram{T}).

Catlab.Theories.terminal — Method

Terminal object.

To implement for a type T, define the method limit(::EmptyDiagram{T}).

Catlab.CategoricalAlgebra.FinSets — Module

Computing in the category of finite sets and functions.

Catlab.CategoricalAlgebra.FinSets.FinFunction — Type

Function between finite sets.

The function can be defined implicitly by an arbitrary Julia function, in which case it is evaluated lazily, or explictly by a vector of integers. In the latter case, the function (1↦1, 2↦3, 3↦2, 4↦3), for example, is represented by the vector [1,3,2,3].

Catlab.CategoricalAlgebra.FinSets.FinFunctionCallable — Type

Function in FinSet defined by a callable Julia object.

To be evaluated lazily unless forced.

Catlab.CategoricalAlgebra.FinSets.FinFunctionVector — Type

Function in FinSet represented explicitly by a vector.

The elements of the set are assumed to be {1,...,n}.

Catlab.CategoricalAlgebra.FinSets.FinSet — Type

Finite set.

This generic type encompasses the category FinSet of finite sets and functions, through types FinSet{S} where S <: AbstractSet, as well as the skeleton of this category, through the type FinSet{Int}. In the latter case, the object FinSet(n) represents the set {1,...,n}.

Catlab.CategoricalAlgebra.FinSets.force — Method

Force evaluation of lazy function or relation.

Catlab.CategoricalAlgebra.FinRelations — Module

Computing in the category of finite sets and relations, and its skeleton.

Catlab.CategoricalAlgebra.FinRelations.BoolRig — Type

The rig of booleans.

This struct is needed because in base Julia, the product of booleans is another boolean, but the sum of booleans is coerced to an integer: true + true == 2.

Catlab.CategoricalAlgebra.FinRelations.FinRel — Type

Object in the category of finite sets and relations.

See also: FinSet.

Catlab.CategoricalAlgebra.FinRelations.FinRelation — Type

Binary relation between finite sets.

A morphism in the category of finite sets and relations. The relation can be represented implicitly by an arbitrary Julia function mapping pairs of elements to booleans or explicitly by a matrix (dense or sparse) taking values in the rig of booleans (BoolRig).

Catlab.CategoricalAlgebra.FinRelations.FinRelationCallable — Type

Relation in FinRel defined by a callable Julia object.

Catlab.CategoricalAlgebra.FinRelations.FinRelationMatrix — Type

Relation in FinRel represented by a boolean matrix.

Boolean matrices are also known as logical matrices or relation matrices.

Catlab.CategoricalAlgebra.Permutations — Module

Computing with permutations: the computer algebra of the symmetric group.

Catlab.CategoricalAlgebra.Permutations.decompose_permutation_by_bubble_sort! — Method

Decompose permutation into adjacent transpositions using bubble sort.

An adjacent transposition, also known as a simple transposition, is a transposition of form (i i+1), represented here as simply the number i.

This algorithm appears as Algorithm 2.7 in the PhD thesis of Jonathan Huang, "Probabilistic reasonsing and learning on permutations: Exploiting structural decompositions of the symmetric group". As Huang notes, the algorithm is very similar to the well-known bubble sort. It has quadratic complexity.

See also: decompose_permutation_by_insertion_sort!

Catlab.CategoricalAlgebra.Permutations.decompose_permutation_by_insertion_sort! — Method

Decompose permutation into adjacent transpositions using insertion sort.

An adjacent transposition, also known as a simple transposition, is a transposition of form (i i+1), represented here as simply the number i.

Bubble sort and insertion sort are, in a sense, dual algorithms (Knuth, TAOCP, Vol 3: Searching and Sort, Sec 5.3.4: Networks for sorting, Figures 45 & 46). A minimal example on which they give different decompositions is the permutation:

[1,2,3] ↦ [3,2,1]

See also: decompose_permutation_by_bubble_sort!

Catlab.CategoricalAlgebra.Permutations.permutation_to_expr — Method

Convert a typed permutation into a morphism expression.

Warning: The morphism expression is not simplified.

Catlab.CategoricalAlgebra.CSets — Module

Categories of C-sets and attributed C-sets.

Catlab.CategoricalAlgebra.CSets.ACSetTransformation — Type

Transformation between attributed C-sets.

A morphism of C-sets is a natural transformation: a transformation between functors C -> Set satisfying the naturality axiom for all morphisms in C. This struct records the data of a transformation; it does not enforce naturality.

The transformation has a component for every object in C. When C-sets have attributes, the data types are assumed to be fixed. Thus, the naturality axiom for data attributes is a commutative triangle, rather than a commutative square.

Catlab.CategoricalAlgebra.CSets.is_natural — Method

Is the transformation between C-sets a natural transformation?

Uses the fact that to check whether a transformation is natural, it suffices to check the naturality equation on a generating set of morphisms.