Categorical algebra

Sets and Relations

The following APIs implement FinSet, the category of Finite Sets (actually the skeleton of FinSet). The objects of this category are natural numbers where n represents a set with n elements. The morphisms are functions between such sets. We use the skeleton of FinSet in order to ensure that all sets are finite and morphisms can be stored using lists of integers. Finite relations are built out of FinSet and can be used to do some relational algebra.

Catlab.CategoricalAlgebra.Sets — Module

Category of (possibly infinite) sets and functions.

This module defines generic types for the category of sets (SetOb, SetFunction), as well as a few basic concrete types, such as a wrapper type to view Julia types as sets (TypeSet). Extensive support for finite sets is provided by another module, FinSets.

Catlab.CategoricalAlgebra.Sets.ConstantFunction — Type

Function in Set taking a constant value.

Catlab.CategoricalAlgebra.Sets.PredicatedSet — Type

Set defined by a predicate (boolean-valued function) on a Julia data type.

Catlab.CategoricalAlgebra.Sets.SetFunction — Type

Abstract type for morphism in the category Set.

Every instance of SetFunction{<:SetOb{T},<:SetOb{T′}} is callable with elements of type T, returning an element of type T′.

Note: This type would be better called simply Function but that name is already taken by the base Julia type.

Catlab.CategoricalAlgebra.Sets.SetOb — Type

Abstract type for object in the category Set.

The type parameter T is the element type of the set.

Note: This type is more abstract than the built-in Julia types AbstractSet and Set, which are intended for data structures for finite sets. Those are encompassed by the subtype FinSet.

Catlab.CategoricalAlgebra.Sets.TypeSet — Type

A Julia data type regarded as a set.

Catlab.Theories.Ob — Method

Forgetful functor Ob: Cat → Set.

Sends a category to its set of objects and a functor to its object map.

Catlab.CategoricalAlgebra.FinSets — Module

The category of finite sets and functions, and its skeleton.

Catlab.CategoricalAlgebra.FinSets.FinDomFunction — Type

Function out of a finite set.

This class of functions is convenient because it is exactly the class that can be represented explicitly by a vector of values from the codomain.

Catlab.CategoricalAlgebra.FinSets.FinFunction — Type

Function between finite sets.

The function can be defined implicitly by an arbitrary Julia function, in which case it is evaluated lazily, or explictly by a vector of integers. In the vector representation, the function (1↦1, 2↦3, 3↦2, 4↦3), for example, is represented by the vector [1,3,2,3].

FinFunctions can be constructed with or without an explicitly provided codomain. If a codomain is provided, by default the constructor checks it is valid.

This type is mildly generalized by FinDomFunction.

Catlab.CategoricalAlgebra.FinSets.FinSet — Type

Finite set.

A finite set has abstract type FinSet{S,T}. The second type parameter T is the element type of the set and the first parameter S is the collection type, which can be a subtype of AbstractSet or another Julia collection type. In addition, the skeleton of the category FinSet is the important special case S = Int. The set ${1,…,n}$ is represented by the object FinSet(n) of type FinSet{Int,Int}.

Catlab.CategoricalAlgebra.FinSets.HashJoin — Type

Hash join algorithm.

Catlab.CategoricalAlgebra.FinSets.JoinAlgorithm — Type

Algorithm for limit of cospan or multicospan with feet being finite sets.

In the context of relational databases, such limits are called joins. The trivial join algorithm is NestedLoopJoin, which is algorithmically equivalent to the generic algorithm ComposeProductEqualizer. The algorithms HashJoin and SortMergeJoin are usually much faster. If you are unsure what algorithm to pick, use SmartJoin.

Catlab.CategoricalAlgebra.FinSets.NestedLoopJoin — Type

Nested-loop join algorithm.

This is the naive algorithm for computing joins.

Catlab.CategoricalAlgebra.FinSets.SmartJoin — Type

Meta-algorithm for joins that attempts to pick an appropriate algorithm.

Catlab.CategoricalAlgebra.FinSets.SortMergeJoin — Type

Sort-merge join algorithm.

Catlab.CategoricalAlgebra.FinSets.SubFinSet — Type

Subset of a finite set.

Catlab.CategoricalAlgebra.FinSets.SubOpBoolean — Type

Algorithm to compute subobject operations using elementwise boolean logic.

Catlab.CategoricalAlgebra.FinSets.TabularLimit — Type

Limit of finite sets viewed as a table.

Any limit of finite sets can be canonically viewed as a table (TabularSet) whose columns are the legs of the limit cone and whose rows correspond to elements of the limit object. To construct this table from an already computed limit, call TabularLimit(::AbstractLimit; ...). The column names of the table are given by the optional argument names.

In this tabular form, applying the universal property of the limit is trivial since it is just tupling. Thus, this representation can be useful when the original limit algorithm does not support efficient application of the universal property. On the other hand, this representation has the disadvantage of generally making the element type of the limit set more complicated.

Catlab.CategoricalAlgebra.FinSets.TabularSet — Type

Finite set whose elements are rows of a table.

The underlying table should be compliant with Tables.jl. For the sake of uniformity, the rows are provided as named tuples, which assumes that the table is not "extremely wide". This should not be a major limitation in practice but see the Tables.jl documentation for further discussion.

Catlab.CategoricalAlgebra.FinSets.VarFunction — Type

Data type of a map out of a set of attribute variables

Currently, domains are FinSet{Int} and codomains are expected to be FinSet{Int}. This could be generalized to being FinSet{Symbol} to allow for symbolic attributes. (Likewise, AttrVars will have to wrap Any rather than Int)

Catlab.CategoricalAlgebra.FinSets.is_indexed — Method

Whether the given function is indexed, i.e., supports efficient preimages.

Catlab.CategoricalAlgebra.FinSets.preimage — Method

The preimage (inverse image) of the value y in the codomain.

Catlab.CategoricalAlgebra.FinRelations — Module

The category of finite sets and relations, and its skeleton.

Catlab.CategoricalAlgebra.FinRelations.BoolRig — Type

The rig of booleans.

This struct is needed because in base Julia, the product of booleans is another boolean, but the sum of booleans is coerced to an integer: true + true == 2.

Catlab.CategoricalAlgebra.FinRelations.FinRel — Type

Object in the category of finite sets and relations.

See also: FinSet.

Catlab.CategoricalAlgebra.FinRelations.FinRelation — Type

Binary relation between finite sets.

A morphism in the category of finite sets and relations. The relation can be represented implicitly by an arbitrary Julia function mapping pairs of elements to booleans or explicitly by a matrix (dense or sparse) taking values in the rig of booleans (BoolRig).

Catlab.CategoricalAlgebra.FinRelations.FinRelationCallable — Type

Relation in FinRel defined by a callable Julia object.

Catlab.CategoricalAlgebra.FinRelations.FinRelationMatrix — Type

Relation in FinRel represented by a boolean matrix.

Boolean matrices are also known as logical matrices or relation matrices.

Free Diagrams, Limits, and Colimts

The following modules define free diagrams in an arbitrary category and specify limit and colimt cones over said diagrams. Thes constructions enjoy the fullest support for FinSet and are used below to define presheaf categories as C-Sets. The general idea of these functions is that you set up a limit computation by specifying a diagram and asking for a limit or colimit cone, which is returned as a struct containing the apex object and the leg morphisms. This cone structure can be queried using the functions apex and legs. Julia's multiple dispatch feature is heavily used to specialize limit and colimit computations for various diagram shapes like product/coproduct and equalizer/coequalizer. As a consumer of this API, it is highly recommended that you use multiple dispatch to specialize your code on the diagram shape whenever possible.

Catlab.CategoricalAlgebra.FreeDiagrams — Module

Free diagrams in a category.

A free diagram in a category is a diagram whose shape is a free category. Examples include the empty diagram, pairs of objects, discrete diagrams, parallel pairs, composable pairs, and spans and cospans. Limits and colimits are most commonly taken over free diagrams.

Catlab.CategoricalAlgebra.FreeDiagrams.BipartiteFreeDiagram — Type

A free diagram with a bipartite structure.

Such diagrams include most of the fixed shapes, such as spans, cospans, and parallel morphisms. They are also the generic shape of diagrams for limits and colimits arising from undirected wiring diagrams. For limits, the boxes correspond to vertices in $V₁$ and the junctions to vertices in $V₂$. Colimits are dual.

Catlab.CategoricalAlgebra.FreeDiagrams.BipartiteFreeDiagram — Method

Convert a free diagram to a bipartite free diagram.

Reduce a free diagram to a free bipartite diagram with the same limit (the default, colimit=false) or the same colimit (colimit=true). The reduction is essentially the same in both cases, except for the choice of where to put isolated vertices, where we follow the conventions described at cone_objects and cocone_objects. The resulting object is a bipartite free diagram equipped with maps from the vertices of the bipartite diagram to the vertices of the original diagram.

Catlab.CategoricalAlgebra.FreeDiagrams.ComposableMorphisms — Type

Composable morphisms in a category.

Composable morphisms are a sequence of morphisms in a category that form a path in the underlying graph of the category.

For the common special case of two morphisms, see ComposablePair.

Catlab.CategoricalAlgebra.FreeDiagrams.ComposablePair — Type

Pair of composable morphisms in a category.

Composable pairs are a common special case of ComposableMorphisms.

Catlab.CategoricalAlgebra.FreeDiagrams.Cospan — Type

Cospan of morphisms in a category.

A common special case of Multicospan. See also Span.

Catlab.CategoricalAlgebra.FreeDiagrams.DiscreteDiagram — Type

Discrete diagram: a diagram with no non-identity morphisms.

Catlab.CategoricalAlgebra.FreeDiagrams.FixedShapeFreeDiagram — Type

Abstract type for free diagram of fixed shape.

Catlab.CategoricalAlgebra.FreeDiagrams.Multicospan — Type

Multicospan of morphisms in a category.

A multicospan is like a Cospan except that it may have a number of legs different than two. A limit of this shape is a pullback.

Catlab.CategoricalAlgebra.FreeDiagrams.Multispan — Type

Multispan of morphisms in a category.

A multispan is like a Span except that it may have a number of legs different than two. A colimit of this shape is a pushout.

Catlab.CategoricalAlgebra.FreeDiagrams.ParallelMorphisms — Type

Parallel morphims in a category.

Parallel morphisms are just morphisms with the same domain and codomain. A (co)limit of this shape is a (co)equalizer.

For the common special case of two morphisms, see ParallelPair.

Catlab.CategoricalAlgebra.FreeDiagrams.ParallelPair — Type

Pair of parallel morphisms in a category.

A common special case of ParallelMorphisms.

Catlab.CategoricalAlgebra.FreeDiagrams.Span — Type

Span of morphims in a category.

A common special case of Multispan. See also Cospan.

Catlab.CategoricalAlgebra.FreeDiagrams.apex — Method

Apex of multispan or multicospan.

The apex of a multi(co)span is the object that is the (co)domain of all the legs.

Catlab.CategoricalAlgebra.FreeDiagrams.bundle_legs — Method

Bundle together legs of a multi(co)span.

For example, calling bundle_legs(span, SVector((1,2),(3,4))) on a multispan with four legs gives a span whose left leg bundles legs 1 and 2 and whose right leg bundles legs 3 and 4. Note that in addition to bundling, this function can also permute legs and discard them.

The bundling is performed using the universal property of (co)products, which assumes that these (co)limits exist.

Catlab.CategoricalAlgebra.FreeDiagrams.cocone_objects — Method

Objects in diagram that will have explicit legs in colimit cocone.

See also: cone_objects.

Catlab.CategoricalAlgebra.FreeDiagrams.cone_objects — Method

Objects in diagram that will have explicit legs in limit cone.

In category theory, it is common practice to elide legs of limit cones that can be computed from other legs, especially for diagrams of certain fixed shapes. For example, when it taking a pullback (the limit of a cospan), the limit object is often treated as having two projections, rather than three. This function encodes such conventions by listing the objects in the diagram that will have corresponding legs in the limit object created by Catlab.

See also: cocone_objects.

Catlab.CategoricalAlgebra.FreeDiagrams.diagram_type — Function

Given a diagram in a category $C$, return Julia type of objects and morphisms in $C$ as a tuple type of form $Tuple{Ob,Hom}$.

Catlab.CategoricalAlgebra.FreeDiagrams.feet — Method

Feet of multispan or multicospan.

The feet of a multispan are the codomains of the legs.

Catlab.CategoricalAlgebra.FreeDiagrams.left — Method

Left leg of span or cospan.

Catlab.CategoricalAlgebra.FreeDiagrams.legs — Method

Legs of multispan or multicospan.

The legs are the morphisms comprising the multi(co)span.

Catlab.CategoricalAlgebra.FreeDiagrams.right — Method

Right leg of span or cospan.

Catlab.CategoricalAlgebra.Limits — Module

Limits and colimits in a category.

Catlab.CategoricalAlgebra.Limits.AbstractColimit — Type

Abstract type for colimit in a category.

The standard concrete subtype is Colimit, although for computational reasons certain categories may use different subtypes to include extra data.

Catlab.CategoricalAlgebra.Limits.AbstractLimit — Type

Abstract type for limit in a category.

The standard concrete subtype is Limit, although for computational reasons certain categories may use different subtypes to include extra data.

Catlab.CategoricalAlgebra.Limits.Colimit — Type

Colimit in a category.

Catlab.CategoricalAlgebra.Limits.ColimitAlgorithm — Type

Algorithm for computing colimits.

Catlab.CategoricalAlgebra.Limits.ComposeCoproductCoequalizer — Type

Compute pushout by composing a coproduct with a coequalizer.

See also: ComposeProductEqualizer.

Catlab.CategoricalAlgebra.Limits.ComposeProductEqualizer — Type

Compute pullback by composing a product with an equalizer.

See also: ComposeCoproductCoequalizer.

Catlab.CategoricalAlgebra.Limits.Limit — Type

Limit in a category.

Catlab.CategoricalAlgebra.Limits.LimitAlgorithm — Type

Algorithm for computing limits.

Catlab.CategoricalAlgebra.Limits.SpecializeColimit — Type

Meta-algorithm that reduces general colimits to common special cases.

Dual to SpecializeLimit.

Catlab.CategoricalAlgebra.Limits.SpecializeLimit — Type

Meta-algorithm that reduces general limits to common special cases.

Reduces limits of free diagrams that happen to be discrete to products. If this fails, fall back to the given algorithm (if any).

TODO: Reduce free diagrams that are (multi)cospans to (wide) pullbacks.

Catlab.CategoricalAlgebra.Limits.ToBipartiteColimit — Type

Compute a colimit by reducing the diagram to a free bipartite diagram.

Catlab.CategoricalAlgebra.Limits.ToBipartiteLimit — Type

Compute a limit by reducing the diagram to a free bipartite diagram.

Catlab.CategoricalAlgebra.Limits.coimage — Method

https://en.wikipedia.org/wiki/Coimage

Catlab.CategoricalAlgebra.Limits.colimit — Method

Colimit of a diagram.

To define colimits in a category with objects Ob, override the method colimit(::FreeDiagram{Ob}) for general colimits or colimit(::D) with suitable type D <: FixedShapeFreeDiagram{Ob} for colimits of specific shape, such as coproducts or coequalizers.

See also: limit

Catlab.CategoricalAlgebra.Limits.epi_mono — Method

The image and coimage are isomorphic. We get this isomorphism using univeral properties.

  CoIm′ ╌╌> I ↠ CoIm
    ┆ ⌟     |
    v       v
    I   ⟶  R ↩ Im
    |       ┆
    v    ⌜  v
    R ╌╌> Im′

Catlab.CategoricalAlgebra.Limits.image — Method

https://en.wikipedia.org/wiki/Image(categorytheory)#Second_definition

Catlab.CategoricalAlgebra.Limits.limit — Method

Limit of a diagram.

To define limits in a category with objects Ob, override the method limit(::FreeDiagram{Ob}) for general limits or limit(::D) with suitable type D <: FixedShapeFreeDiagram{Ob} for limits of specific shape, such as products or equalizers.

See also: colimit

Catlab.CategoricalAlgebra.Limits.pullback — Method

Pullback of a pair of morphisms with common codomain.

To implement for a type T, define the method limit(::Cospan{T}) and/or limit(::Multicospan{T}) or, if you have already implemented products and equalizers, rely on the default implementation.

Catlab.CategoricalAlgebra.Limits.pushout — Method

Pushout of a pair of morphisms with common domain.

To implement for a type T, define the method colimit(::Span{T}) and/or colimit(::Multispan{T}) or, if you have already implemented coproducts and coequalizers, rely on the default implementation.

Catlab.Theories.coequalizer — Method

Coequalizer of morphisms with common domain and codomain.

To implement for a type T, define the method colimit(::ParallelPair{T}) or colimit(::ParallelMorphisms{T}).

Catlab.Theories.copair — Method

Copairing of morphisms: universal property of coproducts/pushouts.

To implement for coproducts of type T, define the method universal(::BinaryCoproduct{T}, ::Cospan{T}) and/or universal(::Coproduct{T}, ::Multicospan{T}) and similarly for pushouts.

Catlab.Theories.coproduct — Method

Coproduct of objects.

To implement for a type T, define the method colimit(::ObjectPair{T}) and/or colimit(::DiscreteDiagram{T}).

Catlab.Theories.create — Method

Unique morphism out of an initial object.

To implement for a type T, define the method universal(::Initial{T}, ::SMulticospan{0,T}).

Catlab.Theories.delete — Method

Unique morphism into a terminal object.

To implement for a type T, define the method universal(::Terminal{T}, ::SMultispan{0,T}).

Catlab.Theories.equalizer — Method

Equalizer of morphisms with common domain and codomain.

To implement for a type T, define the method limit(::ParallelPair{T}) and/or limit(::ParallelMorphisms{T}).

Catlab.Theories.factorize — Method

Factor morphism through (co)equalizer, via the universal property.

To implement for equalizers of type T, define the method universal(::Equalizer{T}, ::SMultispan{1,T}). For coequalizers of type T, define the method universal(::Coequalizer{T}, ::SMulticospan{1,T}).

Catlab.Theories.initial — Method

Initial object.

To implement for a type T, define the method colimit(::EmptyDiagram{T}).

Catlab.Theories.pair — Method

Pairing of morphisms: universal property of products/pullbacks.

To implement for products of type T, define the method universal(::BinaryProduct{T}, ::Span{T}) and/or universal(::Product{T}, ::Multispan{T}) and similarly for pullbacks.

Catlab.Theories.product — Method

Product of objects.

To implement for a type T, define the method limit(::ObjectPair{T}) and/or limit(::DiscreteDiagram{T}).

Catlab.Theories.terminal — Method

Terminal object.

To implement for a type T, define the method limit(::EmptyDiagram{T}).

Catlab.Theories.universal — Function

Universal property of (co)limits.

Compute the morphism whose existence and uniqueness is guaranteed by the universal property of (co)limits.

See also: limit, colimit.

Catlab.CategoricalAlgebra.Limits.@cartesian_monoidal_instance — Macro

Define cartesian monoidal structure using limits.

Implements an instance of ThCartesianCategory assuming that finite products have been implemented following the limits interface.

Catlab.CategoricalAlgebra.Limits.@cocartesian_monoidal_instance — Macro

Define cocartesian monoidal structure using colimits.

Implements an instance of ThCocartesianCategory assuming that finite coproducts have been implemented following the colimits interface.

Categories

Catlab.CategoricalAlgebra.Categories — Module

2-category of categories, functors, and natural transformations.

Categories in mathematics appear in the large, often as categories of sets with extra structure, and in the small, as algebraic structures that generalize groups, monoids, preorders, and graphs. This division manifests in Catlab as well. Large categories (in spirit, if not in the technical sense) occur throughout Catlab as @instances of the theory of categories. For computational reasons, small categories are usually presented by generators and relations.

This module provides a minimal interface to accomodate both situations. Category instances are supported through the wrapper type TypeCat. Finitely presented categories are provided by another module, FinCats.

Catlab.CategoricalAlgebra.Categories.Cat — Type

Alias for Category.

Catlab.CategoricalAlgebra.Categories.CatSize — Type

Size of a category, used for dispatch and subtyping purposes.

A Category type having a particular CatSize means that categories of that type are at most that large.

Catlab.CategoricalAlgebra.Categories.Category — Type

Abstract base type for a category.

The objects and morphisms in the category have Julia types Ob and Hom, respectively. Note that these types do not necessarily form an @instance of the theory of categories, as they may not meaningfully form a category outside the context of this object. For example, a finite category regarded as a reflexive graph with a composition operation might have type Cat{Int,Int}, where the objects and morphisms are numerical identifiers for vertices and edges in the graph.

The basic operations available in any category are: dom, codom, id, compose.

Catlab.CategoricalAlgebra.Categories.CompositeFunctor — Type

Composite of functors.

Catlab.CategoricalAlgebra.Categories.Functor — Type

Abstract base type for a functor between categories.

A functor has a domain and a codomain (dom and codom), which are categories, and object and morphism maps, which can be evaluated using ob_map and hom_map. The functor object can also be called directly when the objects and morphisms have distinct Julia types. This is sometimes but not always the case (see Category), so when writing generic code one should prefer the ob_map and hom_map functions.

Catlab.CategoricalAlgebra.Categories.FunctorCallable — Type

Functor defined by two Julia callables, an object map and a morphism map.

Catlab.CategoricalAlgebra.Categories.IdentityFunctor — Type

Identity functor on a category.

Catlab.CategoricalAlgebra.Categories.LargeCatSize — Type

Size of a large category, such as Set.

To the extent that they form a category, we regard Julia types and functions (TypeCat) as forming a large category.

Catlab.CategoricalAlgebra.Categories.OppositeCat — Type

Opposite category, where morphism are reversed.

Call op(::Cat) instead of directly instantiating this type.

Catlab.CategoricalAlgebra.Categories.OppositeFunctor — Type

Opposite functor, given by the same mapping between opposite categories.

Call op(::Functor) instead of directly instantiating this type.

Catlab.CategoricalAlgebra.Categories.Transformation — Type

Abstract base type for a natural transformation between functors.

A natural transformation $α: F ⇒ G$ has a domain $F$ and codomain $G$ (dom and codom), which are functors $F,G: C → D$ having the same domain $C$ and codomain $D$. The transformation consists of a component $αₓ: Fx → Gx$ in $D$ for each object $x ∈ C$, accessible using component or indexing notation (Base.getindex).

Catlab.CategoricalAlgebra.Categories.TypeCat — Type

Pair of Julia types regarded as a category.

The Julia types should form an @instance of the theory of categories (Theories.Category).

ACSets.Schemas.codom — Method

Codomain of morphism in category.

ACSets.Schemas.dom — Method

Domain of morphism in category.

ACSets.Schemas.hom — Function

Coerce or look up morphism in category.

See also: ob.

ACSets.Schemas.ob — Function

Coerce or look up object in category.

Converts the input to an object in the category, which should be of type Ob in a category of type Cat{Ob,Hom}. How this works depends on the category, but a common case is to look up objects, which might be integers or GAT expressions, by their human-readable name, usually a symbol.

See also: hom.

Catlab.CategoricalAlgebra.Categories.co — Function

2-cell dual of a 2-category.

Catlab.CategoricalAlgebra.Categories.codom_ob — Method

Codomain object of natural transformation.

Given $α: F ⇒ G: C → D$, this function returns $D$.

Catlab.CategoricalAlgebra.Categories.component — Function

Component of natural transformation.

Catlab.CategoricalAlgebra.Categories.dom_ob — Method

Domain object of natural transformation.

Given $α: F ⇒ G: C → D$, this function returns $C$.

Catlab.CategoricalAlgebra.Categories.hom_map — Method

Evaluate functor on morphism.

Catlab.CategoricalAlgebra.Categories.is_hom_equal — Method

Are two morphisms in a category equal?

By default, just checks for equality of Julia objects using $==$. In some categories, checking equality of morphisms may involve nontrivial reasoning.

Catlab.CategoricalAlgebra.Categories.ob_map — Method

Evaluate functor on object.

Catlab.CategoricalAlgebra.Categories.op — Method

Oppositization 2-functor.

The oppositization endo-2-functor on Cat, sending a category to its opposite, is covariant on objects and morphisms and contravariant on 2-morphisms, i.e., is a 2-functor $op: Catᶜᵒ → Cat$. For more explanation, see the nLab.

Catlab.Theories.compose — Method

Compose morphisms in a category.

Catlab.Theories.id — Method

Identity morphism on object in category.

Catlab.CategoricalAlgebra.FinCats — Module

2-category of finitely presented categories.

This module is for the 2-category Cat what the module FinSets is for the category Set: a finitary, combinatorial setting where explicit calculations can be carried out. We emphasize that the prefix Fin means "finitely presented," not "finite," as finite categories are too restrictive a notion for many purposes. For example, the free category on a graph is finite if and only if the graph is DAG, which is a fairly special condition. This usage of Fin is also consistent with FinSet because for sets, being finite and being finitely presented are equivalent.

Catlab.CategoricalAlgebra.FinCats.FinCat — Type

A finitely presented (but not necessarily finite!) category.

Catlab.CategoricalAlgebra.FinCats.FinCatGraph — Type

Abstract type for category backed by finite generating graph.

Catlab.CategoricalAlgebra.FinCats.FinCatGraphEq — Type

Category presented by a finite graph together with path equations.

The objects of the category are vertices in the graph and the morphisms are paths, quotiented by the congruence relation generated by the path equations. See (Spivak, 2014, Category theory for the sciences, 4.5).

Catlab.CategoricalAlgebra.FinCats.FinCatPathGraph — Type

Abstract type for category whose morphisms are paths in a graph.

(Or equivalence classes of paths in a graph, but we compute with

Catlab.CategoricalAlgebra.FinCats.FinCatPresentation — Type

Category defined by a Presentation object.

The presentation type can, of course, be a category (Theories.Category). It can also be a schema (Theories.Schema). In this case, the schema's objects and attribute types are regarded as the category's objects and the schema's morphisms, attributes, and attribute types as the category's morphisms (where the attribute types are identity morphisms). When the schema is formalized as a profunctor whose codomain category is discrete, this amounts to taking the collage of the profunctor.

Catlab.CategoricalAlgebra.FinCats.FinCatSize — Type

Size of a finitely presented category.

Catlab.CategoricalAlgebra.FinCats.FinDomFunctor — Type

A functor out of a finitely presented category.

Catlab.CategoricalAlgebra.FinCats.FinDomFunctorMap — Type

Functor out of a finitely presented category given by explicit mappings.

Catlab.CategoricalAlgebra.FinCats.FinFunctor — Type

A functor between finitely presented categories.

Catlab.CategoricalAlgebra.FinCats.FinTransformation — Type

A natural transformation whose domain category is finitely generated.

This type is for natural transformations $α: F ⇒ G: C → D$ such that the domain category $C$ is finitely generated. Such a natural transformation is given by a finite amount of data (one morphism in $D$ for each generating object of $C$) and its naturality is verified by finitely many equations (one equation for each generating morphism of $C$).

Catlab.CategoricalAlgebra.FinCats.FinTransformationMap — Type

Natural transformation with components given by explicit mapping.

Catlab.CategoricalAlgebra.FinCats.FreeCatGraph — Type

Free category generated by a finite graph.

The objects of the free category are vertices in the graph and the morphisms are (possibly empty) paths.

Catlab.CategoricalAlgebra.FinCats.Path — Type

Path in a graph.

The path is allowed to be empty but always has definite start and end points (source and target vertices).

Catlab.CategoricalAlgebra.FinCats.collect_hom — Method

Collect assignments of functor's morphism map as a vector.

Catlab.CategoricalAlgebra.FinCats.collect_ob — Method

Collect assignments of functor's object map as a vector.

Catlab.CategoricalAlgebra.FinCats.components — Method

Components of a natural transformation.

Catlab.CategoricalAlgebra.FinCats.force — Function

Force evaluation of lazily defined function or functor.

Catlab.CategoricalAlgebra.FinCats.functoriality_failures — Method

Is the purported functor on a presented category functorial?

This function checks that functor preserves domains and codomains. When check_equations is true (the default is false), it also purports to check that the functor preserves all equations in the domain category. No nontrivial checks are currently implemented, so this only functions for identity functors.

See also: is_natural.

Catlab.CategoricalAlgebra.FinCats.graph — Method

Generating graph for a finitely presented category.

Catlab.CategoricalAlgebra.FinCats.hom_generator — Function

Coerce or look up morphism generator in a finitely presented category.

Since morphism generators often have a different data type than morphisms (e.g., in a free category on a graph, the morphism generators are edges and the morphisms are paths), the return type of this function is generally different than that of hom.

Catlab.CategoricalAlgebra.FinCats.hom_generator_name — Function

Name of morphism generator, if any.

When morphism generators have names, this function is a one-sided inverse to hom_generator. See also: ob_generator_name.

Catlab.CategoricalAlgebra.FinCats.hom_generators — Function

Morphism generators of finitely presented category.

Catlab.CategoricalAlgebra.FinCats.is_discrete — Method

Is the category discrete?

A category is discrete if it is has no non-identity morphisms.

Catlab.CategoricalAlgebra.FinCats.is_free — Method

Is the category freely generated?

Catlab.CategoricalAlgebra.FinCats.is_initial — Method

Dual to a final functor, an initial functor is one for which pulling back diagrams along it does not change the limits of these diagrams.

This amounts to checking, for a functor C->D, that, for every object d in Ob(D), the comma category (F/d) is connected.

Catlab.CategoricalAlgebra.FinCats.is_natural — Method

Is the transformation between FinDomFunctors a natural transformation?

This function uses the fact that to check whether a transformation is natural, it suffices to check the naturality equations on a generating set of morphisms of the domain category. In some cases, checking the equations may be expensive or impossible. When the keyword argument check_equations is false, only the domains and codomains of the components are checked.

See also: is_functorial.

Catlab.CategoricalAlgebra.FinCats.ob_generator — Function

Coerce or look up object generator in a finitely presented category.

Because object generators usually coincide with objects, the default method for ob in finitely presented categories simply calls this function.

Catlab.CategoricalAlgebra.FinCats.ob_generator_name — Function

Name of object generator, if any.

When object generators have names, this function is a one-sided inverse to ob_generator in that ob_generator(C, ob_generator_name(C, x)) == x.

Catlab.CategoricalAlgebra.FinCats.ob_generators — Function

Object generators of finitely presented category.

The object generators of finite presented category are almost always the same as the objects. In principle, however, it is possible to have equations between objects, so that there are fewer objects than object generators.

Acsets

Overview and Theory

For an in depth look into the theory behind acsets, we refer the reader to our paper on the subject, which also gives some sense as to how the implementation works. Here, we give a brief tutorial before the the API documentation.

The most essential part of the acset machinery is the schema. The schema parameterizes the acset: each schema corresponds to a different type of acset. Schemas consist of four parts.

Objects $X,Y$ ((X,Y,Z)::Ob)

Homomorphisms $f \colon X \to Y$ (f :: Hom(X,Y)), which go from objects to objects

Attribute types $\mathtt{T}$ (T :: AttrType)

Attributes $a \colon X \to \mathtt{T}$ (a :: Attr(X,T)), which go from objects to data types

For those with a categorical background, a schema is a presentation of a category $|S|$ along with a functor $S$ from $|S|$ to the arrow category $0 \to 1$, such that $S^{-1}(1)$ is discrete.

An acset $F$ on a schema consists of...

a set $F(X)$ of "primary keys" for each object

a function $F(f) \colon F(X) \to F(Y)$ for each morphism

a Julia data type $F(\mathtt{T})$ for each data type $\mathtt{T}$

a function $F(a) \colon F(X) \to F(\mathtt{T})$ for each attribute $a$.

For those with a categorical background, an acset on a schema $S$ consists of a functor from $S$ to $\mathsf{Set}$, such that objects in $S^{-1}(0)$ map to finite sets, and objects in $S^{-1}(1)$ map to sets that represent types. For any particular functor $K \colon S^{-1}(1) \to \mathsf{Set}$, we can also take the category of acsets that restrict to this map on $S^{-1}$.

We can also add equations to this presentation, but we currently do nothing with those equations in the implementation; they mostly serve as documentation.

We will now give an example of how this all works in practice.

using Catlab, Catlab.CategoricalAlgebra

# Write down the schema for a weighted graph
@present SchWeightedGraph(FreeSchema) begin
  V::Ob
  E::Ob
  src::Hom(E,V)
  tgt::Hom(E,V)
  T::AttrType
  weight::Attr(E,T)
end

# Construct the type used to store acsets on the previous schema
# We *index* src and tgt, which means that we store not only
# the forwards map, but also the backwards map.
@acset_type WeightedGraph(SchWeightedGraph, index=[:src,:tgt])

# Construct a weighted graph, with floats as edge weights
g = @acset WeightedGraph{Float64} begin
  V = 4
  E = 5
  src = [1,1,1,2,3]
  tgt = [2,3,4,4,4]
  weight = [7.2, 9.3, 9.4, 0.1, 42.0]
end

Main.__atexample__334.WeightedGraph{Float64} with elements V = 1:4, E = 1:5, T = 1:0

E 	src 	tgt 	weight

1 	1 	2 	7.2

2 	1 	3 	9.3

3 	1 	4 	9.4

4 	2 	4 	0.1

5 	3 	4 	42.0

API

The mathematical abstraction of an acset can of course be implemented in many different ways. Currently, there are three implementations of acsets in Catlab, which share a great deal of code.

These implementations can be split into two categories.

The first category is static acset types. In this implementation, different schemas correspond to different Julia types. Methods on these Julia types are then custom-generated for the schema, using CompTime.jl.

Under this category, there are two classes of static acset types. The first class is acset types that are generated using the @acset_type macro. These acset types are custom-derived structs. The advantage of this is that the structs have names like Graph or WiringDiagram that are printed out in error messages. The disadvantage is that if you are taking in schemas at runtime, you have to eval code in order to use them.

Here is an example of using @acset_type

@acset_type WeightedGraph(SchWeightedGraph, index=[:src,:tgt])
g = WeightedGraph()

The second class is AnonACSets. Like acset types derived from @acset_type, these contain the schema in their type. However, they also contain the type of their fields in their types, so the types printed out in error messages are long and ugly. The advantage of these is that they can be used in situations where the schema is passed in at runtime, and they don't require using eval to create a new acset type.

Here is an example of using AnonACSet

const WeightedGraph = AnonACSetType(SchWeightedGraph, index=[:src,:tgt])
g = WeightedGraph()

The second category is dynamic acset types. Currently, there is just one type that falls under this category: DynamicACSet. This type has a field for the schema, and no code-generation is done for operations on acsets of this type. This means that if the schema is large compared to the data, this type will often be faster than the static acsets.

However, dynamics acsets are a new addition to Catlab, and much of the machinery of limits, colimits, and other high-level acset constructions assumes that the schema of an acset can be derived from the type. Thus, more work will have to be done before dynamic acsets become a drop-in replacement for static acsets.

Here is an example of using a dynamic acset

g = DynamicACSet("WeightedGraph", SchWeightedGraph; index=[:src,:tgt])

Catlab.CategoricalAlgebra.CSets — Module

Categories of C-sets and attributed C-sets.

Catlab.CategoricalAlgebra.CSets.ACSetHomomorphismAlgorithm — Type

Algorithm for finding homomorphisms between attributed $C$-sets.

Catlab.CategoricalAlgebra.CSets.ACSetTransformation — Type

Transformation between attributed C-sets.

Homomorphisms of attributed C-sets generalize homomorphisms of C-sets (CSetTransformation), which you should understand before reading this.

A homomorphism of attributed C-sets with schema S: C ↛ A (a profunctor) is a natural transformation between the corresponding functors col(S) → Set, where col(S) is the collage of S. When the components on attribute types, indexed by objects of A, are all identity functions, the morphism is called tight; in general, it is called loose. With this terminology, acsets on a fixed schema are the objects of an ℳ-category (see Catlab.Theories.MCategory). Calling ACSetTransformation will construct a tight or loose morphism as appropriate, depending on which components are specified.

Since every tight morphism can be considered a loose one, the distinction between tight and loose may seem a minor technicality, but it has important consequences because limits and colimits in a category depend as much on the morphisms as on the objects. In particular, limits and colimits of acsets differ greatly depending on whether they are taken in the category of acsets with tight morphisms or with loose morphisms. Tight morphisms suffice for many purposes, including most applications of colimits. However, when computing limits of acsets, loose morphisms are usually preferable. For more information about limits and colimits in these categories, see TightACSetTransformation and LooseACSetTransformation.

Catlab.CategoricalAlgebra.CSets.ACSetTransformation — Method

Move components as first argument

Catlab.CategoricalAlgebra.CSets.ACSetTransformation — Method

A map f (from A to B) as a map from A to a subobject of B

i.e. get the image of f as a subobject of B

Catlab.CategoricalAlgebra.CSets.ACSetTransformation — Method

A map f (from A to B) as a map of subobjects of A to subjects of B

Catlab.CategoricalAlgebra.CSets.BacktrackingSearch — Type

Find attributed $C$-set homomorphisms using backtracking search.

This procedure uses the classic backtracking search algorithm for a combinatorial constraint satisfaction problem (CSP). As is well known, the homomorphism problem for relational databases is reducible to CSP. Since the C-set homomorphism problem is "the same" as the database homomorphism problem (insofar as attributed C-sets are "the same" as relational databases), it is also reducible to CSP. Backtracking search for CSP is described in many computer science textbooks, such as (Russell & Norvig 2010, Artificial Intelligence, Third Ed., Chapter 6: Constraint satisfaction problems, esp. Algorithm 6.5). In our implementation, the search tree is ordered using the popular heuristic of "minimum remaining values" (MRV), also known as "most constrained variable.

Catlab.CategoricalAlgebra.CSets.CSetTransformation — Type

Transformation between C-sets.

Recall that a C-set homomorphism is a natural transformation: a transformation between functors C → Set satisfying the naturality axiom for every morphism, or equivalently every generating morphism, in C.

This data type records the data of a C-set transformation. Naturality is not strictly enforced but is expected to be satisfied. It can be checked using the function is_natural.

Catlab.CategoricalAlgebra.CSets.HomomorphismQuery — Type

Find attributed $C$-set homomorphisms using a conjunctive query.

This algorithm evaluates a conjunctive query (limit in FinSet) to find all homomorphisms between two $C$-sets. In fact, conjunctive queries are exactly the representable functors from $C$-sets to sets, so every conjunctive query arises in this way, with the caveat that conjunctive queries may correspond to to infinite $C$-sets when $C$ is infinite (but possibly finitely presented).

Catlab.CategoricalAlgebra.CSets.LooseACSetTransformation — Type

Loose transformation between attributed C-sets.

Limits and colimits in the category of attributed C-sets and loose homomorphisms are computed pointwise on both objects and attribute types. This implies that (co)limits of Julia types must be computed. Due to limitations in the expressivity of Julia's type system, only certain simple kinds of (co)limits, such as products, are supported.

Alternatively, colimits involving loose acset transformations can be constructed with respect to explicitly given attribute type components for the legs of the cocone, via the keyword argument type_components to colimit and related functions. This uses the universal property of the colimit. To see how this works, notice that a diagram of acsets and loose acset transformations can be expressed as a diagram D: J → C-Set (for the C-sets) along with another diagram A: J → C-Set (for the attribute sets) and a natural transformation α: D ⇒ A (assigning attributes). Given a natural transformation τ: A ⇒ ΔB to a constant functor ΔB, with components given by type_components, the composite transformation α⋅τ: D ⇒ ΔB is a cocone under D, hence factors through the colimit cocone of D. This factoring yields an assigment of attributes to the colimit in C-Set.

For the distinction between tight and loose, see ACSetTranformation.

Catlab.CategoricalAlgebra.CSets.SubCSet — Type

Sub-C-set of a C-set.

Catlab.CategoricalAlgebra.CSets.TightACSetTransformation — Type

Tight transformation between attributed C-sets.

The category of attributed C-sets and tight homomorphisms is isomorphic to a slice category of C-Set, as explained in our paper "Categorical Data Structures for Technical Computing". Colimits in this category thus reduce to colimits of C-sets, by a standard result about slice categories. Limits are more complicated and are currently not supported.

For the distinction between tight and loose, see ACSetTranformation.

ACSets.ACSetInterface.copy_parts! — Method

Copy parts from a set-valued FinDomFunctor to an ACSet.

Catlab.CategoricalAlgebra.CSets.homomorphism — Method

Find a homomorphism between two attributed $C$-sets.

Returns nothing if no homomorphism exists. For many categories $C$, the $C$-set homomorphism problem is NP-complete and thus this procedure generally runs in exponential time. It works best when the domain object is small.

To restrict to monomorphisms, or homomorphisms whose components are all injective functions, set the keyword argument monic=true. To restrict only certain components to be injective or bijective, use monic=[...] or iso=[...]. For example, setting monic=[:V] for a graph homomorphism ensures that the vertex map is injective but imposes no constraints on the edge map.

To restrict the homomorphism to a given partial assignment, set the keyword argument initial. For example, to fix the first source vertex to the third target vertex in a graph homomorphism, set initial=(V=Dict(1 => 3),). Use the keyword argument type_components to specify nontrivial components on attribute types for a loose homomorphism. These components must be callable: either Julia functions on the appropriate types or FinFunction(Dict(...)).

Use the keyword argument alg to set the homomorphism-finding algorithm. By default, a backtracking search algorithm is used (BacktrackingSearch). Use the keyword argument error_failures = true to get errors explaining any immediate inconsistencies in specified initial data.

See also: homomorphisms, isomorphism.

Catlab.CategoricalAlgebra.CSets.homomorphisms — Method

Find all homomorphisms between two attributed $C$-sets.

This function is at least as expensive as homomorphism and when no homomorphisms exist, it is exactly as expensive.

Catlab.CategoricalAlgebra.CSets.is_homomorphic — Method

Is the first attributed $C$-set homomorphic to the second?

This function generally reduces to homomorphism but certain algorithms may have minor optimizations.

Catlab.CategoricalAlgebra.CSets.is_isomorphic — Method

Are the two attributed $C$-sets isomorphic?

This function generally reduces to isomorphism but certain algorithms may have minor optimizations.

Catlab.CategoricalAlgebra.CSets.isomorphism — Method

Find an isomorphism between two attributed $C$-sets, if one exists.

See homomorphism for more information about the algorithms involved.

Catlab.CategoricalAlgebra.CSets.isomorphisms — Method

Find all isomorphisms between two attributed $C$-sets.

This function is at least as expensive as isomorphism and when no homomorphisms exist, it is exactly as expensive.

Catlab.CategoricalAlgebra.CSets.maximum_common_subobject — Method

Compute the Maximimum Common C-Sets from a vector of C-Sets.

Find an ACSet `$a$ with maximum possible size ($|a|$) such that there is a monic span of ACSets $a₁ ← a → a₂$. There may be many maps from this overlap into each of the inputs, so a Vector{Vector{ACSetTransformations}} per overlap is returned (a cartesian product can be taken of these to obtain all possible multispans for that overlap). If there are multiple overlaps of equal size, these are all returned.

If there are attributes, we ignore these and use variables in the apex of the overlap.

Catlab.CategoricalAlgebra.CSets.naturality_failures — Method

Returns a dictionary whose keys are contained in the names in arrows(S) and whose value at :f, for an arrow (f,c,d), is a lazy iterator over the elements of X(c) on which α's naturality square for f does not commute. Components should be a NamedTuple or Dictionary with keys contained in the names of S's morphisms and values vectors or dicts defining partial functions from X(c) to Y(c).

Catlab.CategoricalAlgebra.CSets.subobject_graph — Method

Preorder of subobjects via inclusion. Returns a graph + list of subobjects corresponding to its vertices. The subobjects are ordered by decreasing size (so it's topologically sorted)

Catlab.CategoricalAlgebra.FinCats.is_natural — Method

Check naturality condition for a purported ACSetTransformation, α: X→Y. For each hom in the schema, e.g. h: m → n, the following square must commute: αₘ Xₘ –> Yₘ Xₕ ↓ ✓ ↓ Yₕ Xₙ –> Yₙ αₙ You're allowed to run this on a named tuple partly specifying an ACSetTransformation, though at this time the domain and codomain must be fully specified ACSets.

Catlab.CategoricalAlgebra.CSets.@acset_transformation — Macro

Provides a shorthand for constructing a tight acset transformation by giving its components. Homomorphism search allows partial specification, with the return value being the unique extension if it exists.

Keyword arguments can be passed on to the search function after the body of the transformation.

Example usage for transformation between WeightedGraph{String}:

@acset_transformation A B begin
  V = [3,5,2] #complete specification can be a vector
  E = Dict(1 => 3, 4 => 3) #otherwise use a dict
end monic=true iso=[:V] or [:V,:E], etc

Catlab.CategoricalAlgebra.DataMigrations — Module

Functorial data migration for attributed C-sets.

Catlab.CategoricalAlgebra.DataMigrations.DataMigration — Type

Data migration functor given contravariantly.

This type encompasses data migration functors from $C$-sets to $D$-sets given contravariantly by a functor out of the schema $D$. The simplest such functor is a pullback data migration (DeltaMigration), specified by a functor $D → C$ between the schemas. Other important cases include conjunctive and duc data migrations.

Catlab.CategoricalAlgebra.DataMigrations.DeltaMigration — Type

Delta or pullback functorial data migration between acsets.

Given a functor $F: D → C$, the delta migration functor $Δ_F$ from $C$-sets to $D$-sets is defined contravariantly by $Δ_F(X) := X ∘ F$.

See (Spivak, 2014, Category Theory for the Sciences) for details.

Catlab.CategoricalAlgebra.DataMigrations.SigmaMigration — Type

Sigma or left pushforward functorial data migration between acsets.

Given a functor $F: C → D$, the sigma data migration $Σ_F$ is a functor from $C$-sets to $D$-sets that is left adjoint to the delta migration functor $Δ_F$ (DeltaMigration). Explicitly, the $D$-set $Σ_F(X)$ is given on objects $d ∈ D$ by the formula $Σ_F(x)(d) = \mathrm{colim}_{F ↓ d} X ∘ π$, where $π: (F ↓ d) → C$ is the projection.

See (Spivak, 2014, Category Theory for the Sciences) for details.

Catlab.CategoricalAlgebra.DataMigrations.SigmaMigration — Method

Create a C-Set for the collage of the functor. Initialize data in the domain portion of the collage, then run the chase.

Catlab.CategoricalAlgebra.DataMigrations.colimit_representables — Method

Interpret conjunctive data migration as a colimit of representables.

Given a conjunctive data migration (a functor J → Diag{op}(C)) and the Yoneda embedding for C (a functor op(C) → C-Set computed via yoneda), take colimits of representables to construct a op(J)-shaped diagram of C-sets.

Since every C-set is a colimit of representables, this is a generic way of constructing diagrams of C-sets.

Catlab.CategoricalAlgebra.DataMigrations.migrate! — Method

Contravariantly migrate data from the acset Y to the acset X.

This is the mutating variant of migrate!. When the functor on schemas is the identity, this operation is equivalent to copy_parts!.

Catlab.CategoricalAlgebra.DataMigrations.migrate — Method

Contravariantly migrate data from the acset Y to a new acset of type T.

The mutating variant of this function is migrate!.

Catlab.CategoricalAlgebra.DataMigrations.representable — Method

Construct a representable C-set.

Recall that a representable C-set is one of form $C(c,-): C → Set$ for some object $c ∈ C$.

This function computes the $c$ representable as the left pushforward data migration of the singleton ${c}$-set along the inclusion functor ${c} ↪ C$, which works because left Kan extensions take representables to representables (Mac Lane 1978, Exercise X.3.2). Besides the intrinsic difficulties with representables (they can be infinite), this function thus inherits any limitations of our implementation of left pushforward data migrations.

Catlab.CategoricalAlgebra.DataMigrations.representable — Method

ACSet types do not store info about equations, so this info is lost when we try to recover the presentation from the datatype. Thus, this method for representable should only be used for free schemas

Catlab.CategoricalAlgebra.DataMigrations.yoneda — Method

Yoneda embedding of category C in category of C-sets.

Because Catlab privileges copresheaves (C-sets) over presheaves, this is the contravariant Yoneda embedding, i.e., the embedding C^op → C-Set.

If representables have already been computed (which can be expensive) they can be provided via the cache keyword argument.

Input cons is a constructor for the ACSet

Returns a FinDomFunctor with domain op(C).

Catlab.CategoricalAlgebra.Chase — Module

The chase is an algorithm which subjects a C-Set instance to constraints expressed in the language of regular logic, called embedded dependencies (EDs, or 'triggers').

A morphism S->T, encodes an embedded dependency. If the pattern S is matched (via a homomorphism S->I), we demand there exist a morphism T->I (for some database instance I) that makes the triangle commute in order to satisfy the dependency (if this is not the case, then the trigger is 'active').

Homomorphisms can merge elements and introduce new ones. The former kind are called "equality generating dependencies" (EGDs) and the latter "tuple generating dependencies" (TGDs). Any homomorphism can be factored into EGD and TGD components by, respectively, restricting the codomain to the image or restricting the domain to the coimage.

Catlab.CategoricalAlgebra.Chase.chase — Method

chase(I::ACSet, Σ::AbstractDict, n::Int)

Chase a C-Set or C-Rel instance given a list of embedded dependencies. This may not terminate, so a bound n on the number of iterations is required.

[,]

ΣS ⟶ Iₙ ⊕↓ ⋮ (resulting morphism) ΣT ... Iₙ₊₁

There is a copy of S and T for each active trigger. A trigger is a map from S into the current instance. What makes it 'active' is that there is no morphism from T to I that makes the triangle commute.

Each iteration constructs the above pushout square. The result is a morphism, so that one can keep track of the provenance of elements in the original CSet instance within the chased result.

Whether or not the result is due to success or timeout is returned as a boolean flag.

TODO: this algorithm could be made more efficient by homomorphism caching.

Catlab.CategoricalAlgebra.StructuredCospans — Module

Structured cospans.

This module provides a generic interface for structured cospans with a concrete implementation for attributed C-sets.

Catlab.CategoricalAlgebra.StructuredCospans.OpenACSetLeg — Type

Leg of a structured (multi)cospan of acsets in R-form.

A convenience type that contains the data of an acset transformation, except for the codomain, since that data is already given by the decoration of the R-form structured cospan.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredCospan — Type

Structured cospan.

The first type parameter L encodes a functor L: A → X from the base category A, often FinSet, to a category X with "extra structure." An L-structured cospan is then a cospan in X whose feet are images under L of objects in A. The category X is assumed to have pushouts.

Structured cospans form a double category with no further assumptions on the functor L. To obtain a symmetric monoidal double category, L must preserve finite coproducts. In practice, L usually has a right adjoint R: X → A, which implies that L preserves all finite colimits. It also allows structured cospans to be constructed more conveniently from an object x in X plus a cospan in A with apex R(x).

See also: StructuredMulticospan.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredCospan — Method

Construct structured cospan in R-form.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredCospan — Method

Construct structured cospan in L-form.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredCospanOb — Type

Object in the category of L-structured cospans.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredMulticospan — Type

Structured multicospan.

A structured multicospan is like a structured cospan except that it may have a number of legs different than two.

See also: StructuredCospan.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredMulticospan — Method

Construct structured multicospan in R-form.

Catlab.CategoricalAlgebra.StructuredCospans.OpenACSetTypes — Method

Create types for open attributed C-sets from an attributed C-set type.

The type parameters of the given acset type should not be instantiated with specific Julia types. This function returns a pair of types, one for objects, a subtype of StructuredCospanOb, and one for morphisms, a subtype of StructuredMulticospan. Both types will have the same type parameters for attribute types as the given acset type.

Mathematically speaking, this function sets up structured (multi)cospans with a functor $L: A → X$ between categories of acsets that creates "discrete acsets." Such a "discrete acset functor" is a functor that is left adjoint to a certain kind of forgetful functor between categories of acsets, namely one that is a pullback along an inclusion of schemas such that the image of inclusion has no outgoing arrows. For example, the schema inclusion ${V} ↪ {E ⇉ V}$ has this property but ${E} ↪ {E ⇉ V}$ does not.

See also: OpenCSetTypes.

Catlab.CategoricalAlgebra.StructuredCospans.OpenCSetTypes — Method

Create types for open C-sets from a C-set type.

A special case of OpenACSetTypes. See there for details.