Categorical algebra

Catlab.CategoricalAlgebra.FreeDiagrams — Module

Free diagrams in a category.

A free diagram in a category is a diagram whose shape is a free category. Examples include the empty diagram, discrete diagrams, parallel morphisms, spans, and cospans. Limits and colimits are most commonly taken over free diagrams.

Catlab.CategoricalAlgebra.FreeDiagrams.Cospan — Type

Cospan of morphisms in a category.

A common special case of Multicospan. See also Span.

Catlab.CategoricalAlgebra.FreeDiagrams.DecoratedCospan — Type

Decorate Cospan of morphisms for representing open networks.

Catlab.CategoricalAlgebra.FreeDiagrams.FixedFreeDiagram — Type

Abstract type for free diagram of fixed shape.

Catlab.CategoricalAlgebra.FreeDiagrams.Multicospan — Type

Multicospan of morphisms in a category.

A multicospan is like a Cospan except that it may have a number of legs different than two. A limit of this shape is a pullback.

Catlab.CategoricalAlgebra.FreeDiagrams.Multispan — Type

Multispan of morphisms in a category.

A multispan is like a Span except that it may have a number of legs different than two. A colimit of this shape is a pushout.

Catlab.CategoricalAlgebra.FreeDiagrams.ParallelMorphisms — Type

Parallel morphims in a category.

Parallel morphisms are just morphisms with the same domain and codomain. A (co)limit of this shape is a (co)equalizer.

For the common special case of two morphisms, see ParallelPair.

Catlab.CategoricalAlgebra.FreeDiagrams.ParallelPair — Type

Pair of parallel morphisms in a category.

A common special case of ParallelMorphisms.

Catlab.CategoricalAlgebra.FreeDiagrams.Span — Type

Span of morphims in a category.

A common special case of Multispan. See also Cospan.

Catlab.CategoricalAlgebra.Limits — Module

Limits and colimits in a category.

Catlab.CategoricalAlgebra.Limits.AbstractColimit — Type

Abstract type for colimit in a category.

The standard concrete subtype is Colimit, although for computational reasons certain categories may use different subtypes to include extra data.

Catlab.CategoricalAlgebra.Limits.AbstractLimit — Type

Abstract type for limit in a category.

The standard concrete subtype is Limit, although for computational reasons certain categories may use different subtypes to include extra data.

Catlab.CategoricalAlgebra.Limits.Colimit — Type

Colimit in a category.

Catlab.CategoricalAlgebra.Limits.Limit — Type

Limit in a category.

Catlab.CategoricalAlgebra.Limits.pullback — Method

Pullback of a pair of morphisms with common codomain.

Catlab.CategoricalAlgebra.Limits.pullback — Method

Pullback of a cospan.

The default implementation computes the pullback from products and equalizers.

Catlab.CategoricalAlgebra.Limits.pushout — Method

Pushout of a pair of morphisms with common domain.

Catlab.CategoricalAlgebra.Limits.pushout — Method

Pushout of a span.

The default implementation computes the pushout from coproducts and coequalizers.

Catlab.Theories.coproduct — Method

Coproduct of a pair of objects.

Catlab.Theories.equalizer — Method

Equalizer of a pair of morphisms with common domain and codomain.

Catlab.Theories.product — Method

Product of a pair of objects.

Catlab.CategoricalAlgebra.FinSets — Module

Computing in the category of finite sets and functions.

Catlab.CategoricalAlgebra.FinSets.FinFunction — Type

Function between finite sets.

The function can be defined implicitly by an arbitrary Julia function, in which case it is evaluated lazily, or explictly by a vector of integers. In the latter case, the function (1↦1, 2↦3, 3↦2, 4↦3), for example, is represented by the vector [1,3,2,3].

Catlab.CategoricalAlgebra.FinSets.FinFunctionCallable — Type

Function in FinSet defined by a callable Julia object.

To be evaluated lazily unless forced.

Catlab.CategoricalAlgebra.FinSets.FinFunctionVector — Type

Function in FinSet represented explicitly by a vector.

The elements of the set are assumed to be {1,...,n}.

Catlab.CategoricalAlgebra.FinSets.FinSet — Type

Finite set.

This generic type encompasses the category FinSet of finite sets and functions, through types FinSet{S} where S <: AbstractSet, as well as the skeleton of this category, through the type FinSet{Int}. In the latter case, the object FinSet(n) represents the set {1,...,n}.

Catlab.CategoricalAlgebra.FinSets.force — Method

Force evaluation of lazy function or relation.

Catlab.CategoricalAlgebra.FinRelations — Module

Computing in the category of finite sets and relations, and its skeleton.

Catlab.CategoricalAlgebra.FinRelations.BoolRig — Type

The rig of booleans.

This struct is needed because in base Julia, the product of booleans is another boolean, but the sum of booleans is coerced to an integer: true + true == 2.

Catlab.CategoricalAlgebra.FinRelations.FinRel — Type

Object in the category of finite sets and relations.

See also: FinSet.

Catlab.CategoricalAlgebra.FinRelations.FinRelation — Type

Binary relation between finite sets.

A morphism in the category of finite sets and relations. The relation can be represented implicitly by an arbitrary Julia function mapping pairs of elements to booleans or explicitly by a matrix (dense or sparse) taking values in the rig of booleans (BoolRig).

Catlab.CategoricalAlgebra.FinRelations.FinRelationCallable — Type

Relation in FinRel defined by a callable Julia object.

Catlab.CategoricalAlgebra.FinRelations.FinRelationMatrix — Type

Relation in FinRel represented by a boolean matrix.

Boolean matrices are also known as logical matrices or relation matrices.

Catlab.CategoricalAlgebra.Permutations — Module

Computing with permutations: the computer algebra of the symmetric group.

Catlab.CategoricalAlgebra.Permutations.decompose_permutation_by_bubble_sort! — Method

Decompose permutation into adjacent transpositions using bubble sort.

An adjacent transposition, also known as a simple transposition, is a transposition of form (i i+1), represented here as simply the number i.

This algorithm appears as Algorithm 2.7 in the PhD thesis of Jonathan Huang, "Probabilistic reasonsing and learning on permutations: Exploiting structural decompositions of the symmetric group". As Huang notes, the algorithm is very similar to the well-known bubble sort. It has quadratic complexity.

See also: decompose_permutation_by_insertion_sort!

Catlab.CategoricalAlgebra.Permutations.decompose_permutation_by_insertion_sort! — Method

Decompose permutation into adjacent transpositions using insertion sort.

An adjacent transposition, also known as a simple transposition, is a transposition of form (i i+1), represented here as simply the number i.

Bubble sort and insertion sort are, in a sense, dual algorithms (Knuth, TAOCP, Vol 3: Searching and Sort, Sec 5.3.4: Networks for sorting, Figures 45 & 46). A minimal example on which they give different decompositions is the permutation:

[1,2,3] ↦ [3,2,1]

See also: decompose_permutation_by_bubble_sort!

Catlab.CategoricalAlgebra.Permutations.permutation_to_expr — Method

Convert a typed permutation into a morphism expression.

Warning: The morphism expression is not simplified.

Catlab.CategoricalAlgebra.CSets.copy_parts! — Method

Copy parts from one C-set into another. All subparts among the selected parts, including any attached data, are preserved. Thus, if the selected parts form a sub-C-set, then the whole sub-C-set is preserved. On the other hand, if the selected parts do not form a sub-C-set, then some copied parts will have undefined subparts.

Catlab.CategoricalAlgebra.CSets.has_subpart — Method

Whether a C-set has a subpart with the given name.

Catlab.CategoricalAlgebra.Graphs — Module

Data structures for graphs, based on C-sets.

Support for graphs, symmetric graphs, and property graphs.

Catlab.CategoricalAlgebra.Graphs.AbstractPropertyGraph — Type

Abstract type for graph with properties.

Concrete types are PropertyGraph and SymmetricPropertyGraph.

Catlab.CategoricalAlgebra.Graphs.PropertyGraph — Type

Graph with properties.

"Property graphs" are graphs with arbitrary named properties on the graph, vertices, and edges. They are intended for applications with a large number of ad-hoc properties. If you have a small number of known properties, it is better and more efficient to create a specialized C-set type using CSetType.

See also: SymmetricPropertyGraph.

Catlab.CategoricalAlgebra.Graphs.SymmetricPropertyGraph — Type

Symmetric graphs with properties.

The edge properties are preserved under the edge involution, so these can be interpreted as "undirected" property (multi)graphs.

See also: PropertyGraph.