Categorical algebra

FinSet and FinRel

The following APIs implement FinSet, the category of Finite Sets (actually the skeleton of FinSet). The objects of this category are natural numbers where n represents a set with n elements. The morphisms are functions between such sets. We use the skeleton of FinSet in order to ensure that all sets are finite and morphisms can be stored using lists of integers. Finite relations are built out of FinSet and can be used to do some relational algebra.

Catlab.CategoricalAlgebra.FinSets — Module

The category of finite sets and functions, and its skeleton.

Catlab.CategoricalAlgebra.FinSets.FinDomFunction — Type

Function out of a finite set.

This class of functions is convenient because it is exactly the class that can be represented explicitly by a vector of values from the codomain.

Catlab.CategoricalAlgebra.FinSets.FinFunction — Type

Function between finite sets.

The function can be defined implicitly by an arbitrary Julia function, in which case it is evaluated lazily, or explictly by a vector of integers. In the vector representation, the function (1↦1, 2↦3, 3↦2, 4↦3), for example, is represented by the vector [1,3,2,3].

This type is mildly generalized by FinDomFunction.

Catlab.CategoricalAlgebra.FinSets.FinSet — Type

Finite set.

This generic type encompasses the category FinSet of finite sets and functions, through types FinSet{S} where S <: AbstractSet, as well as the skeleton of this category, through the type FinSet{Int}. In the latter case, the object FinSet(n) represents the set ${1,...,n}$.

Catlab.CategoricalAlgebra.FinSets.HashJoin — Type

Hash join algorithm.

Catlab.CategoricalAlgebra.FinSets.JoinAlgorithm — Type

Algorithm for limit of spans or multispans out of finite sets.

In the context of relational databases, such limits are joins.

Catlab.CategoricalAlgebra.FinSets.NestedLoopJoin — Type

Nested-loop join algorithm.

This is the naive algorithm for computing joins.

Catlab.CategoricalAlgebra.FinSets.SmartJoin — Type

Meta-algorithm for joins that attempts to pick an appropriate algorithm.

Catlab.CategoricalAlgebra.FinSets.SortMergeJoin — Type

Sort-merge join algorithm.

Catlab.CategoricalAlgebra.FinSets.force — Method

Force evaluation of lazy function or relation.

Catlab.CategoricalAlgebra.FinSets.is_indexed — Method

Whether the given function is indexed, i.e., supports efficient preimages.

Catlab.CategoricalAlgebra.FinSets.preimage — Method

The preimage (inverse image) of the value y in the codomain.

Catlab.CategoricalAlgebra.FinRelations — Module

The category of finite sets and relations, and its skeleton.

Catlab.CategoricalAlgebra.FinRelations.BoolRig — Type

The rig of booleans.

This struct is needed because in base Julia, the product of booleans is another boolean, but the sum of booleans is coerced to an integer: true + true == 2.

Catlab.CategoricalAlgebra.FinRelations.FinRel — Type

Object in the category of finite sets and relations.

See also: FinSet.

Catlab.CategoricalAlgebra.FinRelations.FinRelation — Type

Binary relation between finite sets.

A morphism in the category of finite sets and relations. The relation can be represented implicitly by an arbitrary Julia function mapping pairs of elements to booleans or explicitly by a matrix (dense or sparse) taking values in the rig of booleans (BoolRig).

Catlab.CategoricalAlgebra.FinRelations.FinRelationCallable — Type

Relation in FinRel defined by a callable Julia object.

Catlab.CategoricalAlgebra.FinRelations.FinRelationMatrix — Type

Relation in FinRel represented by a boolean matrix.

Boolean matrices are also known as logical matrices or relation matrices.

Diagrams, Limits, and Colimts

The following modules define diagrams in an arbitrary category and specify limit and colimt cones over said diagrams. Thes constructions enjoy the fullest support for FinSet and are used below to define presheaf categories as C-Sets. The general idea of these functions is that you set up a limit computation by specifying a diagram and asking for a limit or colimit cone, which is returned as a struct containing the apex object and the leg morphisms. This cone structure can be queried using the functions apex and legs. Julia's multiple dispatch feature is heavily used to specialize limit and colimit computations for various diagram shapes like product/coproduct and equalizer/coequalizer. As a consumer of this API, it is highly recommended that you use multiple dispatch to specialize your code on the diagram shape whenever possible.

Catlab.CategoricalAlgebra.FreeDiagrams — Module

Free diagrams in a category.

A free diagram in a category is a diagram whose shape is a free category. Examples include the empty diagram, pairs of objects, discrete diagrams, parallel morphisms, spans, and cospans. Limits and colimits are most commonly taken over free diagrams.

Catlab.CategoricalAlgebra.FreeDiagrams.BipartiteFreeDiagram — Type

A free diagram that is bipartite.

Such diagrams include most of the fixed shapes, such as spans, cospans, and parallel morphisms. They are the generic shape of diagrams for limits and colimits arising from undirected wiring diagrams. For limits, the boxes correspond to vertices in $V₁$ and the junctions to vertics in $V₂$. Colimits are dual.

Catlab.CategoricalAlgebra.FreeDiagrams.Cospan — Type

Cospan of morphisms in a category.

A common special case of Multicospan. See also Span.

Catlab.CategoricalAlgebra.FreeDiagrams.DiscreteDiagram — Type

Discrete diagram: a diagram whose only morphisms are identities.

Catlab.CategoricalAlgebra.FreeDiagrams.FixedShapeFreeDiagram — Type

Abstract type for free diagram of fixed shape.

Catlab.CategoricalAlgebra.FreeDiagrams.Multicospan — Type

Multicospan of morphisms in a category.

A multicospan is like a Cospan except that it may have a number of legs different than two. A limit of this shape is a pullback.

Catlab.CategoricalAlgebra.FreeDiagrams.Multispan — Type

Multispan of morphisms in a category.

A multispan is like a Span except that it may have a number of legs different than two. A colimit of this shape is a pushout.

Catlab.CategoricalAlgebra.FreeDiagrams.ParallelMorphisms — Type

Parallel morphims in a category.

Parallel morphisms are just morphisms with the same domain and codomain. A (co)limit of this shape is a (co)equalizer.

For the common special case of two morphisms, see ParallelPair.

Catlab.CategoricalAlgebra.FreeDiagrams.ParallelPair — Type

Pair of parallel morphisms in a category.

A common special case of ParallelMorphisms.

Catlab.CategoricalAlgebra.FreeDiagrams.Span — Type

Span of morphims in a category.

A common special case of Multispan. See also Cospan.

Catlab.CategoricalAlgebra.FreeDiagrams.apex — Method

apex(span::Multispan)

returns the object at the top of the multispan, which is the domain of all the legs.

Catlab.CategoricalAlgebra.FreeDiagrams.bundle_legs — Method

Bundle together legs of a multi(co)span.

For example, calling bundle_legs(span, SVector((1,2),(3,4))) on a multispan with four legs gives a span whose left leg bundles legs 1 and 2 and whose right leg bundles legs 3 and 4. Note that in addition to bundling, this function can also permute legs and discard them.

The bundling is performed using the universal property of (co)products, which assumes that these (co)limits exist.

Catlab.CategoricalAlgebra.FreeDiagrams.feet — Method

feet(span::Multispan)

returns the collection of feet in the multspan, which are the codomains of the legs.

Catlab.CategoricalAlgebra.FreeDiagrams.legs — Method

legs(span::Multispan)

returns the collection of legs in the multspan, which are the morphisms sharing a common domain.

Catlab.CategoricalAlgebra.Limits — Module

Limits and colimits in a category.

Catlab.CategoricalAlgebra.Limits.AbstractColimit — Type

Abstract type for colimit in a category.

The standard concrete subtype is Colimit, although for computational reasons certain categories may use different subtypes to include extra data.

Catlab.CategoricalAlgebra.Limits.AbstractLimit — Type

Abstract type for limit in a category.

The standard concrete subtype is Limit, although for computational reasons certain categories may use different subtypes to include extra data.

Catlab.CategoricalAlgebra.Limits.Colimit — Type

Colimit in a category.

Catlab.CategoricalAlgebra.Limits.ColimitAlgorithm — Type

Algorithm for computing colimits.

Catlab.CategoricalAlgebra.Limits.ComposeCoproductCoequalizer — Type

Compute pushout by composing a coproduct with a coequalizer.

See also: ComposeProductEqualizer.

Catlab.CategoricalAlgebra.Limits.ComposeProductEqualizer — Type

Compute pullback by composing a product with an equalizer.

See also: ComposeCoproductCoequalizer.

Catlab.CategoricalAlgebra.Limits.Limit — Type

Limit in a category.

Catlab.CategoricalAlgebra.Limits.LimitAlgorithm — Type

Algorithm for computing limits.

Catlab.CategoricalAlgebra.Limits.colimit — Function

Colimit of a diagram.

To define colimits in a category with objects Ob, override the method colimit(::FreeDiagram{Ob}) for general colimits or colimit(::D) with suitable type D <: FixedShapeFreeDiagram{Ob} for colimits of specific shape, such as coproducts or coequalizers.

See also: limit

Catlab.CategoricalAlgebra.Limits.limit — Function

Limit of a diagram.

To define limits in a category with objects Ob, override the method limit(::FreeDiagram{Ob}) for general limits or limit(::D) with suitable type D <: FixedShapeFreeDiagram{Ob} for limits of specific shape, such as products or equalizers.

See also: colimit

Catlab.CategoricalAlgebra.Limits.pullback — Method

Pullback of a pair of morphisms with common codomain.

To implement for a type T, define the method limit(::Cospan{T}) and/or limit(::Multicospan{T}) or, if you have already implemented products and equalizers, rely on the default implementation.

Catlab.CategoricalAlgebra.Limits.pushout — Method

Pushout of a pair of morphisms with common domain.

To implement for a type T, define the method colimit(::Span{T}) and/or colimit(::Multispan{T}) or, if you have already implemented coproducts and coequalizers, rely on the default implementation.

Catlab.CategoricalAlgebra.Limits.universal — Function

Universal property of (co)limits.

Compute the morphism whose existence and uniqueness is guaranteed by the universal property of (co)limits.

See also: limit, colimit.

Catlab.Theories.coequalizer — Method

Coequalizer of morphisms with common domain and codomain.

To implement for a type T, define the method colimit(::ParallelPair{T}) or colimit(::ParallelMorphisms{T}).

Catlab.Theories.copair — Method

Copairing of morphisms: universal property of coproducts/pushouts.

To implement for coproducts of type T, define the method universal(::BinaryCoproduct{T}, ::Cospan{T}) and/or universal(::Coproduct{T}, ::Multicospan{T}) and similarly for pushouts.

Catlab.Theories.coproduct — Method

Coproduct of objects.

To implement for a type T, define the method colimit(::ObjectPair{T}) and/or colimit(::DiscreteDiagram{T}).

Catlab.Theories.create — Method

Unique morphism out of an initial object.

To implement for a type T, define the method universal(::Initial{T}, ::SMulticospan{0,T}).

Catlab.Theories.delete — Method

Unique morphism into a terminal object.

To implement for a type T, define the method universal(::Terminal{T}, ::SMultispan{0,T}).

Catlab.Theories.equalizer — Method

Equalizer of morphisms with common domain and codomain.

To implement for a type T, define the method limit(::ParallelPair{T}) and/or limit(::ParallelMorphisms{T}).

Catlab.Theories.factorize — Method

Factor morphism through (co)equalizer, via the universal property.

To implement for equalizers of type T, define the method universal(::Equalizer{T}, ::SMultispan{1,T}). For coequalizers of type T, define the method universal(::Coequalizer{T}, ::SMulticospan{1,T}).

Catlab.Theories.initial — Method

Initial object.

To implement for a type T, define the method colimit(::EmptyDiagram{T}).

Catlab.Theories.pair — Method

Pairing of morphisms: universal property of products/pullbacks.

To implement for products of type T, define the method universal(::BinaryProduct{T}, ::Span{T}) and/or universal(::Product{T}, ::Multispan{T}) and similarly for pullbacks.

Catlab.Theories.product — Method

Product of objects.

To implement for a type T, define the method limit(::ObjectPair{T}) and/or limit(::DiscreteDiagram{T}).

Catlab.Theories.terminal — Method

Terminal object.

To implement for a type T, define the method limit(::EmptyDiagram{T}).

Key components of the CSet and ACSet machinery

FreeSchema A finite presentation of a category that will be used as the schema of a database in the algebraic databases conception of categorical database theory. Functors out of a schema into FinSet are combinatorial structures over the schema. Attributes in a schema allow you to encode numerical (any julia type) into the database. You can find several examples of schemas in Catlab.Graphs where they define categorical versions of graph theory.

CSet/AttributedCSet is a struct/constructors whose values (tables, indices) are parameterized by a CatDesc/AttrDesc. These are in memory databases over the schema equiped with ACSetTranformations as natural transformations that encode relationships between database instances.

CSetType/AttributedCSetTypeprovides a function to construct a julia type for ACSet instances, parameterized by CatDesc/AttrDesc. This function constructs the new type at runtime. In order to have the interactive nature of Julia, and to dynamically construct schemas based on runtime values, we need to define new Julia types at runtime. This function converts the schema spec to the corresponding Julia type.

CatDesc/AttrDesc the encoding of a schema into a Julia type. These exist because Julia only allows certain kinds of data in the parameter of a dependent type. Thus, we have to serialize a schema into those primitive data types so that we can use them to parameterize the ACSet type over the schema. This is an implementation detail subject to complete overhaul.

Catlab.CategoricalAlgebra.CSets — Module

Categories of C-sets and attributed C-sets.

Catlab.CategoricalAlgebra.CSets.ACSetTransformation — Type

Transformation between attributed C-sets.

A homomorphism of C-sets is a natural transformation: a transformation between functors C → Set satisfying the naturality axiom for all morphisms in C. This struct records the data of a transformation; it does not enforce naturality, but see is_natural.

A C-set transformation has a component for every object in C. When C-sets have attributes, the data types are assumed to be fixed. Thus, the naturality axiom for data attributes is a commutative triangle, rather than a commutative square.

Catlab.CategoricalAlgebra.CSets.generate_json_acset — Method

Serialize an ACSet object to a JSON string

Catlab.CategoricalAlgebra.CSets.homomorphism — Method

Find a homomorphism between two attributed $C$-sets.

Returns nothing if no homomorphism exists. For many categories $C$, the $C$-set homomorphism problem is NP-complete and thus this procedure generally runs in exponential time. It works best when the domain object is small.

This procedure uses the classic backtracking search algorithm for a combinatorial constraint satisfaction problem (CSP). As is well known, the homomorphism problem for relational databases is equivalent to CSP. Since the C-set homomorphism problem is "the same" as the database homomorphism problem (insofar as attributed C-sets are "the same" as relational databases), it is also equivalent to CSP. Backtracking search for CSP is described in many computer science textbooks, such as (Russell & Norvig 2010, Artificial Intelligence, Third Ed., Chapter 6: Constraint satisfaction problems, esp. Algorithm 6.5). In our implementation, the search tree is ordered using the popular heuristic of "minimum remaining values" (MRV), also known as "most constrained variable."

Set the keyword argument monic=true to restrict to monomorphisms: homomorphisms whose components are all injective functions.

See also: homomorphisms, isomorphism.

Catlab.CategoricalAlgebra.CSets.homomorphisms — Method

Find all homomorphisms between two attributed $C$-sets.

This function is at least as expensive as homomorphism and when no homomorphisms exist, it is exactly as expensive.

Catlab.CategoricalAlgebra.CSets.is_homomorphic — Method

Is the first attributed $C$-set homomorphic to the second?

A convenience function based on homomorphism.

Catlab.CategoricalAlgebra.CSets.is_isomorphic — Method

Are the two attributed $C$-sets isomorphic?

A convenience function based on isomorphism.

Catlab.CategoricalAlgebra.CSets.is_natural — Method

Is the transformation between C-sets a natural transformation?

Uses the fact that to check whether a transformation is natural, it suffices to check the naturality equation on a generating set of morphisms.

Catlab.CategoricalAlgebra.CSets.isomorphism — Method

Find an isomorphism between two attributed $C$-sets, if one exists.

See homomorphism for more information about the algorithms involved.

Catlab.CategoricalAlgebra.CSets.isomorphisms — Method

Find all isomorphisms between two attributed $C$-sets.

This function is at least as expensive as isomorphism and when no homomorphisms exist, it is exactly as expensive.

Catlab.CategoricalAlgebra.CSets.parse_json_acset — Method

Deserialize a dictionary from a parsed JSON string to an object of the given ACSet type

Catlab.CategoricalAlgebra.CSets.parse_json_acset — Method

Deserialize a JSON string to an object of the given ACSet type

Catlab.CategoricalAlgebra.CSets.read_json_acset — Method

Read a JSON file to an object of the given ACSet type

Catlab.CategoricalAlgebra.CSets.write_json_acset — Method

Serialize an ACSet object to a JSON file

Catlab.CategoricalAlgebra.StructuredCospans — Module

Structured cospans.

This module provides a generic interface for structured cospans with a concrete implementation for attributed C-sets.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredCospan — Type

Structured cospan.

The first type parameter L encodes a functor L: A → X from the base category A, often FinSet, to a category X with "extra structure." An L-structured cospan is then a cospan in X whose feet are images under L of objects in A. The category X is assumed to have pushouts.

Structured cospans form a double category with no further assumptions on the functor L. To obtain a symmetric monoidal double category, L must preserve finite coproducts. In practice, L usually has a right adjoint R: X → A, which implies that L preserves all finite colimits. It also allows structured cospans to be constructed more conveniently from an object x in X plus a cospan in A with apex R(x).

See also: StructuredMulticospan.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredCospan — Method

Construct structured cospan in R-form.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredCospan — Method

Construct structured cospan in L-form.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredCospanOb — Type

Object in the category of L-structured cospans.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredMulticospan — Type

Structured multicospan.

A structured multicospan is like a structured cospan except that it may have a number of legs different than two.

See also: StructuredCospan.

Catlab.CategoricalAlgebra.StructuredCospans.StructuredMulticospan — Method

Construct structured multicospan in R-form.

Catlab.CategoricalAlgebra.StructuredCospans.OpenACSetTypes — Method

Create types for open attributed C-sets from an attributed C-set type.

The resulting types, for objects and morphisms, each have the same type parameters for data types as the original type.

See also: OpenCSetTypes.

Catlab.CategoricalAlgebra.StructuredCospans.OpenCSetTypes — Method

Create types for open C-sets from a C-set type.

Returns two types, for objects, a subtype of StructuredCospanOb, and for morphisms, a subtype of StructuredMulticospan.

See also: OpenACSetTypes.