Doctrines

Catlab.Doctrines.AbelianBicategoryRelations — Module

Doctrine of abelian bicategory of relations

Unlike Carboni & Walters, we use additive notation and nomenclature.

References:

Carboni & Walters, 1987, "Cartesian bicategories I", Sec. 5

Baez & Erbele, 2015, "Categories in control"

Catlab.Doctrines.BicategoryRelations — Module

Doctrine of bicategory of relations

TODO: The 2-morphisms are missing. I haven't decided how to handle them yet.

References:

Carboni & Walters, 1987, "Cartesian bicategories I"

Walters, 2009, blog post, "Categorical algebras of relations", http://rfcwalters.blogspot.com/2009/10/categorical-algebras-of-relations.html

Catlab.Doctrines.BiproductCategory — Module

Doctrine of bicategory category

Also known as a semiadditive category.

FIXME: This signature should extend MonoidalCategoryWithBidiagonals, CartesianCategory, and CocartesianCategory, but multiple inheritance is not yet supported.

Catlab.Doctrines.CartesianCategory — Module

Doctrine of cartesian category

Actually, this is a cartesian symmetric monoidal category but we omit these qualifiers for brevity.

Catlab.Doctrines.CartesianClosedCategory — Module

Doctrine of cartesian closed category (aka, CCC)

A CCC is a cartesian category with internal homs (aka, exponential objects).

Catlab.Doctrines.Category — Module

Doctrine of category (with no extra structure)

Warning: We compose functions from left to right, i.e., if f:A→B and g:B→C then compose(f,g):A→C. Under this convention function are applied on the right, e.g., if a∈A then af∈B.

We retain the usual meaning of the symbol ∘ (\circ), i.e., g∘f = compose(f,g). This usage is too entrenched to overturn, inconvenient though it may be. We use symbol ⋅ (\cdot) for diagrammatic composition: f⋅g = compose(f,g).

Catlab.Doctrines.Category2 — Module

Doctrine of (strict) 2-category

Catlab.Doctrines.CocartesianCategory — Module

Doctrine of cocartesian category

Actually, this is a cocartesian symmetric monoidal category but we omit these qualifiers for brevity.

Catlab.Doctrines.CompactClosedCategory — Module

Doctrine of compact closed category

Catlab.Doctrines.DaggerCategory — Module

Doctrine of dagger category

Catlab.Doctrines.DaggerCompactCategory — Module

Doctrine of dagger compact category

In a dagger compact category, there are two kinds of adjoints of a morphism f::Hom(A,B), the adjoint mate mate(f)::Hom(dual(B),dual(A)) and the dagger adjoint dagger(f)::Hom(B,A). In the category of Hilbert spaces, these are respectively the Banach space adjoint and the Hilbert space adjoint (Reed-Simon, Vol I, Sec VI.2). In Julia, they would correspond to transpose and adjoint in the official LinearAlegbra module. For the general relationship between mates and daggers, see Selinger's survey of graphical languages for monoidal categories.

FIXME: This signature should extend both DaggerCategory and CompactClosedCategory, but multiple inheritance is not yet supported.

Catlab.Doctrines.DaggerSymmetricMonoidalCategory — Module

Doctrine of dagger symmetric monoidal category

Also known as a symmetric monoidal dagger category.

FIXME: This signature should extend both DaggerCategory and SymmetricMonoidalCategory, but multiple inheritance is not yet supported.

Catlab.Doctrines.FreeCartesianCategory — Module

Syntax for a free cartesian category.

In this syntax, the pairing and projection operations are defined using duplication and deletion, and do not have their own syntactic elements. Of course, this convention could be reversed.

Catlab.Doctrines.FreeCartesianClosedCategory — Module

Syntax for a free cartesian closed category.

See also FreeCartesianCategory.

Catlab.Doctrines.FreeCategory2 — Module

Syntax for a 2-category.

Checks domains of morphisms but not 2-morphisms.

Catlab.Doctrines.FreeCocartesianCategory — Module

Syntax for a free cocartesian category.

In this syntax, the copairing and inclusion operations are defined using merging and creation, and do not have their own syntactic elements. Of course, this convention could be reversed.

Catlab.Doctrines.MonoidalCategory — Module

Doctrine of monoidal category

To avoid associators and unitors, we assume the monoidal category is strict. By the coherence theorem there is no loss of generality, but we may add a signature for weak monoidal categories later.

Catlab.Doctrines.MonoidalCategoryWithBidiagonals — Module

Doctrine of monoidal category with bidiagonals

The terminology is nonstandard (is there any standard terminology?) but is intended to mean a monoidal category with coherent diagonals and codiagonals. Unlike in a biproduct category, the naturality axioms need not be satisfied.

FIXME: This signature should extend both MonoidalCategoryWithDiagonals and MonoidalCategoryWithCodiagonals, but multiple inheritance is not yet supported.

Catlab.Doctrines.MonoidalCategoryWithCodiagonals — Module

Doctrine of monoidal category with codiagonals

A monoidal category with codiagonals is a symmetric monoidal category equipped with coherent collections of merging and creating morphisms (monoids). Unlike in a cocartesian category, the naturality axioms need not be satisfied.

For references, see MonoidalCategoryWithDiagonals.

Catlab.Doctrines.MonoidalCategoryWithDiagonals — Module

Doctrine of monoidal category with diagonals

A monoidal category with diagonals is a symmetric monoidal category equipped with coherent collections of copying and deleting morphisms (comonoids). Unlike in a cartesian category, the naturality axioms need not be satisfied.

References:

Selinger, 2010, "A survey of graphical languages for monoidal categories", Section 6.6: "Cartesian center"

Selinger, 1999, "Categorical structure of asynchrony"

Catlab.Doctrines.SymmetricMonoidalCategory — Module

Doctrine of symmetric monoidal category

The signature (but not the axioms) is the same as a braided monoidal category.

Base.collect — Method

Collect generators of object in monoidal category as a vector.

Base.ndims — Method

Number of "dimensions" of object in monoidal category.

Catlab.Doctrines.AbelianBicategoryRelations.Hom — Type

Morphism in a category

Catlab.Doctrines.AbelianBicategoryRelations.Ob — Type

Object in a category

Catlab.Doctrines.BicategoryRelations.Hom — Type

Morphism in a category

Catlab.Doctrines.BicategoryRelations.Ob — Type

Object in a category

Catlab.Doctrines.BiproductCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.BiproductCategory.Ob — Type

Object in a category

Catlab.Doctrines.CartesianCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.CartesianCategory.Ob — Type

Object in a category

Catlab.Doctrines.CartesianClosedCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.CartesianClosedCategory.Ob — Type

Object in a category

Catlab.Doctrines.Category.Hom — Type

Morphism in a category

Catlab.Doctrines.Category.Ob — Type

Object in a category

Catlab.Doctrines.Category2.Hom — Type

Morphism in a category

Catlab.Doctrines.Category2.Hom2 — Type

2-morphism in a 2-category

Catlab.Doctrines.Category2.Ob — Type

Object in a category

Catlab.Doctrines.CocartesianCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.CocartesianCategory.Ob — Type

Object in a category

Catlab.Doctrines.CompactClosedCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.CompactClosedCategory.Ob — Type

Object in a category

Catlab.Doctrines.DaggerCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.DaggerCategory.Ob — Type

Object in a category

Catlab.Doctrines.DaggerCompactCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.DaggerCompactCategory.Ob — Type

Object in a category

Catlab.Doctrines.FreeAbelianBicategoryRelations.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeAbelianBicategoryRelations.Ob — Type

Object in a category

Catlab.Doctrines.FreeBicategoryRelations.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeBicategoryRelations.Ob — Type

Object in a category

Catlab.Doctrines.FreeBiproductCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeBiproductCategory.Ob — Type

Object in a category

Catlab.Doctrines.FreeCartesianCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeCartesianCategory.Ob — Type

Object in a category

Catlab.Doctrines.FreeCartesianClosedCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeCartesianClosedCategory.Ob — Type

Object in a category

Catlab.Doctrines.FreeCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeCategory.Ob — Type

Object in a category

Catlab.Doctrines.FreeCategory2.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeCategory2.Hom2 — Type

2-morphism in a 2-category

Catlab.Doctrines.FreeCategory2.Ob — Type

Object in a category

Catlab.Doctrines.FreeCocartesianCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeCocartesianCategory.Ob — Type

Object in a category

Catlab.Doctrines.FreeCompactClosedCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeCompactClosedCategory.Ob — Type

Object in a category

Catlab.Doctrines.FreeDaggerCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeDaggerCategory.Ob — Type

Object in a category

Catlab.Doctrines.FreeDaggerCompactCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeDaggerCompactCategory.Ob — Type

Object in a category

Catlab.Doctrines.FreeSymmetricMonoidalCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.FreeSymmetricMonoidalCategory.Ob — Type

Object in a category

Catlab.Doctrines.MonoidalCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.MonoidalCategory.Ob — Type

Object in a category

Catlab.Doctrines.MonoidalCategoryWithBidiagonals.Hom — Type

Morphism in a category

Catlab.Doctrines.MonoidalCategoryWithBidiagonals.Ob — Type

Object in a category

Catlab.Doctrines.MonoidalCategoryWithCodiagonals.Hom — Type

Morphism in a category

Catlab.Doctrines.MonoidalCategoryWithCodiagonals.Ob — Type

Object in a category

Catlab.Doctrines.MonoidalCategoryWithDiagonals.Hom — Type

Morphism in a category

Catlab.Doctrines.MonoidalCategoryWithDiagonals.Ob — Type

Object in a category

Catlab.Doctrines.SymmetricMonoidalCategory.Hom — Type

Morphism in a category

Catlab.Doctrines.SymmetricMonoidalCategory.Ob — Type

Object in a category