Graphs

Catlab.Graphs.BasicGraphs — Module

Data structures for graphs, based on C-sets.

Provides the category theorist's four basic kinds of graphs: graphs (aka directed multigraphs), symmetric graphs, reflexive graphs, and symmetric reflexive graphs. Also defines half-edge graphs. The API generally follows that of LightGraphs.jl, with some departures due to differences between the data structures.

Catlab.Graphs.BasicGraphs.AbstractGraph — Type

Abstract type for graphs, aka directed multigraphs.

Catlab.Graphs.BasicGraphs.AbstractHalfEdgeGraph — Type

Abstract type for half-edge graphs, possibly with data attributes.

Catlab.Graphs.BasicGraphs.AbstractReflexiveGraph — Type

Abstract type for reflexive graphs, possibly with data attributes.

Catlab.Graphs.BasicGraphs.AbstractSymmetricGraph — Type

Abstract type for symmetric graph, possibly with data attributes.

Catlab.Graphs.BasicGraphs.AbstractSymmetricReflexiveGraph — Type

Abstract type for symmetric reflexive graphs, possibly with data attributes.

Catlab.Graphs.BasicGraphs.AbstractSymmetricWeightedGraph — Type

Abstract type for symmetric weights graphs.

Catlab.Graphs.BasicGraphs.AbstractWeightedGraph — Type

Abstract type for weighted graphs.

Catlab.Graphs.BasicGraphs.HasGraph — Type

Abstract type for C-sets that contain a graph.

This type encompasses C-sets where the schema for graphs is a subcategory of C. This includes, for example, graphs, symmetric graphs, and reflexive graphs, but not half-edge graphs.

Catlab.Graphs.BasicGraphs.HasVertices — Type

Abstract type for C-sets that contain vertices.

This type encompasses C-sets where the schema C contains an object V interpreted as vertices. This includes, for example, graphs and half-edge graphs, but not bipartite graphs or wiring diagrams.

Base.inv — Method

Involution on half-edge(s) in a half-edge graph.

Base.inv — Method

Involution on edge(s) in a symmetric graph.

Catlab.Graphs.BasicGraphs.add_dangling_edge! — Method

Add a dangling edge to a half-edge graph.

A "dangling edge" is a half-edge that is paired with itself under the half-edge involution. They are usually interpreted differently than "self-loops", i.e., a pair of distinct half-edges incident to the same vertex.

Catlab.Graphs.BasicGraphs.add_dangling_edges! — Method

Add multiple dangling edges to a half-edge graph.

Catlab.Graphs.BasicGraphs.add_edge! — Method

Add an edge to a graph.

Catlab.Graphs.BasicGraphs.add_edges! — Method

Add multiple edges to a graph.

Catlab.Graphs.BasicGraphs.add_vertex! — Method

Add a vertex to a graph.

Catlab.Graphs.BasicGraphs.add_vertices! — Method

Add multiple vertices to a graph.

Catlab.Graphs.BasicGraphs.all_neighbors — Method

Union of in-neighbors and out-neighbors in a graph.

Catlab.Graphs.BasicGraphs.edges — Method

Edges in a graph, or between two vertices in a graph.

Catlab.Graphs.BasicGraphs.half_edges — Method

Half-edges in a half-edge graph, or incident to a vertex.

Catlab.Graphs.BasicGraphs.has_edge — Method

Whether the graph has the given edge, or an edge between two vertices.

Catlab.Graphs.BasicGraphs.has_vertex — Method

Whether the graph has the given vertex.

Catlab.Graphs.BasicGraphs.induced_subgraph — Method

Subgraph induced by a set of a vertices.

The induced subgraph consists of the given vertices and all edges between vertices in this set.

Catlab.Graphs.BasicGraphs.inneighbors — Method

In-neighbors of vertex in a graph.

Catlab.Graphs.BasicGraphs.ne — Method

Number of edges in a graph, or between two vertices in a graph.

In a symmetric graph, this function counts both edges in each edge pair, so that the number of edges in a symmetric graph is twice the number of edges in the corresponding undirected graph (at least when the edge involution has no fixed points).

Catlab.Graphs.BasicGraphs.neighbors — Method

Neighbors of vertex in a graph.

In a graph, this function is an alias for outneighbors; in a symmetric graph, a vertex has the same out-neighbors and as in-neighbors, so the distinction is moot.

In the presence of multiple edges, neighboring vertices are given with multiplicity. To get the unique neighbors, call unique(neighbors(g)).

Catlab.Graphs.BasicGraphs.nv — Method

Number of vertices in a graph.

Catlab.Graphs.BasicGraphs.outneighbors — Method

Out-neighbors of vertex in a graph.

Catlab.Graphs.BasicGraphs.refl — Method

Reflexive loop(s) of vertex (vertices) in a reflexive graph.

Catlab.Graphs.BasicGraphs.rem_edge! — Method

Remove an edge from a graph.

Catlab.Graphs.BasicGraphs.rem_edges! — Method

Remove multiple edges from a graph.

Catlab.Graphs.BasicGraphs.rem_vertex! — Method

Remove a vertex from a graph.

When keep_edges is false (the default), all edges incident to the vertex are also deleted. When keep_edges is true, incident edges are preserved but their source/target vertices become undefined.

Catlab.Graphs.BasicGraphs.rem_vertices! — Method

Remove multiple vertices from a graph.

Edges incident to any of the vertices are treated as in rem_vertex!.

Catlab.Graphs.BasicGraphs.src — Method

Source vertex (vertices) of edges(s) in a graph.

Catlab.Graphs.BasicGraphs.tgt — Method

Target vertex (vertices) of edges(s) in a graph.

Catlab.Graphs.BasicGraphs.vertex — Method

Incident vertex (vertices) of half-edge(s) in a half-edge graph.

Catlab.Graphs.BasicGraphs.vertices — Method

Vertices in a graph.

Catlab.Graphs.BasicGraphs.weight — Method

Weight(s) of edge(s) in a weighted graph.

Catlab.Graphs.PropertyGraphs.AbstractPropertyGraph — Type

Abstract type for graph with properties.

Concrete types are PropertyGraph and SymmetricPropertyGraph.

Catlab.Graphs.PropertyGraphs.PropertyGraph — Type

Graph with properties.

"Property graphs" are graphs with arbitrary named properties on the graph, vertices, and edges. They are intended for applications with a large number of ad-hoc properties. If you have a small number of known properties, it is better and more efficient to create a specialized C-set type using CSetType.

See also: SymmetricPropertyGraph.

Catlab.Graphs.PropertyGraphs.SymmetricPropertyGraph — Type

Symmetric graphs with properties.

The edge properties are preserved under the edge involution, so these can be interpreted as "undirected" property (multi)graphs.

See also: PropertyGraph.

Catlab.Graphs.PropertyGraphs.eprops — Method

Properties of edge in a property graph.

Catlab.Graphs.PropertyGraphs.get_eprop — Method

Get property of edge or edges in a property graph.

Catlab.Graphs.PropertyGraphs.get_gprop — Method

Get graph-level property of a property graph.

Catlab.Graphs.PropertyGraphs.get_vprop — Method

Get property of vertex or vertices in a property graph.

Catlab.Graphs.PropertyGraphs.gprops — Method

Graph-level properties of a property graph.

Catlab.Graphs.PropertyGraphs.set_eprop! — Method

Set property of edge or edges in a property graph.

Catlab.Graphs.PropertyGraphs.set_eprops! — Method

Set multiple properties of an edge in a property graph.

Catlab.Graphs.PropertyGraphs.set_gprop! — Method

Set graph-level property in a property graph.

Catlab.Graphs.PropertyGraphs.set_gprops! — Method

Set multiple graph-level properties in a property graph.

Catlab.Graphs.PropertyGraphs.set_vprop! — Method

Set property of vertex or vertices in a property graph.

Catlab.Graphs.PropertyGraphs.set_vprops! — Method

Set multiple properties of a vertex in a property graph.

Catlab.Graphs.PropertyGraphs.vprops — Method

Properties of vertex in a property graph.

Catlab.Graphs.GraphAlgorithms — Module

Algorithms on graphs based on C-sets.

Catlab.Graphs.GraphAlgorithms.connected_component_projection — Function

Projection onto (weakly) connected components of a graph.

Returns a function in FinSet{Int} from the vertex set to the set of components.

Catlab.Graphs.GraphAlgorithms.connected_components — Method

(Weakly) connected components of a graph.

Returns a vector of vectors, which are the components of the graph.

Catlab.Graphs.GraphAlgorithms.topological_sort — Method

Topological sort of a directed acyclic graph.

The depth-first search algorithm is adapted from the function topological_sort_by_dfs in LightGraphs.jl.

Catlab.Graphs.GraphAlgorithms.transitive_reduction! — Method

Transitive reduction of a DAG.

The algorithm computes the longest paths in the DAGs and keeps only the edges corresponding to longest paths of length 1. Requires a topological sort, which is computed if it is not supplied.