Programs

Catlab.Programs.GenerateJuliaPrograms — Module

Compile or evaluate morphisms as Julia programs.

Catlab.Programs.GenerateJuliaPrograms.Block — Type

A block of Julia code with input and output variables.

Catlab.Programs.GenerateJuliaPrograms.CompileState — Type

Internal state for compilation of morphism into Julia code.

Catlab.Programs.GenerateJuliaPrograms.compile — Method

Compile a morphism expression into a Julia function.

Catlab.Programs.GenerateJuliaPrograms.compile_block — Method

Compile a morphism expression into a block of Julia code.

Catlab.Programs.GenerateJuliaPrograms.compile_expr — Method

Compile a morphism expression into a Julia function expression.

Catlab.Programs.GenerateJuliaPrograms.evaluate — Method

Evaluate a morphism as a function.

If the morphism will be evaluated only once (possibly with vectorized inputs), then direct evaluation will be much faster than compiling (via compile) and evaluating a standard Julia function.

Compare with functor.

Catlab.Programs.ParseJuliaPrograms — Module

Parse Julia programs into morphisms represented as wiring diagrams.

Catlab.Programs.ParseJuliaPrograms.parse_wiring_diagram — Method

Parse a wiring diagram from a Julia function expression.

The macro version of this function is @program.

Catlab.Programs.ParseJuliaPrograms.@program — Macro

Parse a wiring diagram from a Julia program.

For the most part, this is standard Julia but we take a few liberties with the syntax. Products are represented as tuples. So if x and y are variables of type $X$ and $Y$, then (x,y) has type $X \otimes Y$. Also, both () and nothing are interpreted as the monoidal unit $I$.

Unlike in standard Julia, the call expressions f(x,y) and f((x,y)) are equivalent. Consequently, given morphisms $f: W \to X \otimes Y$ and $g: X \otimes Y \to Z$, the code

x, y = f(w)
g(x,y)

is equivalent to g(f(w)). In standard Julia, at most one of these calls to g would be valid.

The diagonals (copying and deleting) are implicit in the Julia syntax: copying is variable reuse and deleting is variable non-use. For the codiagonals (merging and creating), a special syntax is provided, reinterpreting Julia's vector literals. The merge of x1 and x2 is represented by the vector [x1,x2] and creation by the empty vector []. For example, f([x1,x2]) translates to compose(mmerge(X),f).

This macro is a wrapper around parse_wiring_diagram.