Wiring diagrams

Catlab.WiringDiagrams.WiringDiagramCore — Module

Generic data structures for wiring diagrams (aka, string diagrams).

A (directed) wiring diagram consists of a collection of boxes with input and output ports connected by wires. A box can be atomic (possessing no internal structure) or can itself be a wiring diagram. Thus, wiring diagrams can be nested recursively. Wiring diagrams are closely related to what the CS literature calls "directed graphs with ports" or more simply "port graphs". The main difference is that a wiring diagram has an "outer box": a wiring diagram has its own ports that can be connected to the ports of its boxes.

This module provides a generic data structure for wiring diagrams. Arbitrary data can be attached to the boxes, ports, and wires of a wiring diagram. The diagrams are "abstract" in the sense that they cannot be directly rendered as raster or vector graphics. However, they form a useful intermediate representation that can be serialized to and from GraphML or translated into Graphviz or other declarative diagram languages.

Catlab.WiringDiagrams.WiringDiagramCore.AbstractBox — Type

Base type for any box (node) in a wiring diagram.

This type represents an arbitrary black box with inputs and outputs.

Catlab.WiringDiagrams.WiringDiagramCore.Box — Type

An atomic box in a wiring diagram.

These boxes have no internal structure.

Catlab.WiringDiagrams.WiringDiagramCore.Port — Type

A port on a box to which wires can be connected.

Catlab.WiringDiagrams.WiringDiagramCore.PortKind — Type

Kind of port: input or output.

Catlab.WiringDiagrams.WiringDiagramCore.Wire — Type

A wire connecting one port to another.

Catlab.WiringDiagrams.WiringDiagramCore.WiringDiagram — Type

A directed wiring diagram, aka a string diagram.

The wiring diagram is implemented using the following internal data structures. A LightGraphs DiGraph stores the "skeleton" of the diagram: a simple directed graph with the boxes as vertices and with an edge between two vertices iff there is at least one wire between the corresponding boxes. There are two special vertices, accessible via input_id and output_id, representing the input and output ports, respectively.

The DiGraph is wrapped inside a MetaDiGraph to attach properties to the vertices and edges. For each edge, an edge property stores the list of wires between the source and target boxes.

Catlab.WiringDiagrams.WiringDiagramCore.encapsulate — Method

Encapsulate multiple boxes within a single sub-diagram.

This operation is a (one-sided) inverse to subsitution (see substitute).

Catlab.WiringDiagrams.WiringDiagramCore.encapsulated_subdiagram — Method

Create an encapsulating box for a set of boxes in a wiring diagram.

To a first approximation, the union of input ports of the given boxes will become the inputs ports of the encapsulating box and likewise for the output ports. However, when copies or merges occur, as in a cartesian or cocartesian category, a simplification procedure may reduce the number of ports on the encapsulating box.

Specifically:

Each input port of an encapsulated box will have at most one incoming wire

from the encapsulating outer box, and each output port of an encapsulated box will have at most one outgoing wire to the encapsulating outer box.

A set of ports connected to the same outside (non-encapsulated) ports will be

simplified into a single port of the encapsulating box.

See also induced_subdiagram.

Catlab.WiringDiagrams.WiringDiagramCore.graph — Method

Retrieve the underlying LightGraphs graph.

Do not mutate it! All mutations should pass through the WiringDiagram methods: add_box!, rem_box!, etc.

Catlab.WiringDiagrams.WiringDiagramCore.in_wires — Method

Get all wires coming into the port.

Catlab.WiringDiagrams.WiringDiagramCore.in_wires — Method

Get all wires coming into the box.

Catlab.WiringDiagrams.WiringDiagramCore.induced_subdiagram — Method

The wiring diagram induced by a subset of its boxes.

See also encapsulated_subdiagram.

Catlab.WiringDiagrams.WiringDiagramCore.is_permuted_equal — Method

Check equality of wiring diagram under permutation of boxes.

When the boxes in the first diagram d1 are permuted according to σ, does it become identical to the second diagram d2?

Catlab.WiringDiagrams.WiringDiagramCore.out_wires — Method

Get all wires coming out of the port.

Catlab.WiringDiagrams.WiringDiagramCore.out_wires — Method

Get all wires coming out of the box.

Catlab.WiringDiagrams.WiringDiagramCore.singleton_diagram — Method

A wiring diagram with a single box connected to its outer ports.

Catlab.WiringDiagrams.WiringDiagramCore.substitute — Method

Substitute wiring diagrams for boxes.

Performs one or more substitutions. When performing multiple substitutions, the substitutions are simultaneous.

This operation implements the operadic composition of wiring diagrams (ocompose).

Catlab.WiringDiagrams.WiringDiagramCore.validate_ports — Method

Check compatibility of source and target ports.

The default implementation is a no-op.

Catlab.WiringDiagrams.WiringDiagramCore.wires — Method

Get all wires coming into or out of the box.

Catlab.WiringDiagrams.WiringLayers — Module

Data structure for connecting one layer to another by wires.

This module defines a generic data structure to represent a wiring between one layer of input ports and another layer of output ports. A wiring layer forms a bipartite graph with independent edge sets the input ports and the output ports.

Wiring layers are an auxillary data structure. They are not very interesting in their own right, but they can be a useful intermediate representation. For example, a morphism expression comprising generators, compositions, products, and wiring layers is intermediate between a pure GAT expression (which has no wiring layers, but may have identities, braidings, copies, etc.) and a wiring diagram, which is purely graphical.

Catlab.WiringDiagrams.WiringLayers.NLayer — Type

Number of input or output ports in a layer.

Object in the category of wiring layers.

Catlab.WiringDiagrams.WiringLayers.WiringLayer — Type

Connection by wires of one layer to another.

Morphism in the category of wiring layers.

Catlab.WiringDiagrams.WiringLayers.complete_layer — Method

Completely connected wiring layer.

The layer's underlying graph is the complete bipartite graph.

Catlab.WiringDiagrams.WiringLayers.to_wiring_diagram — Method

Convert a wiring layer into a wiring diagram.

Catlab.WiringDiagrams.WiringLayers.wiring_layer_between — Method

Wiring layer representing the wires between two boxes in a wiring diagram.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams — Module

Wiring diagrams as a symmetric monoidal category and as an operad.

This module provides a high-level functional and algebraic interface to wiring diagrams, building on the low-level imperative interface. It also defines data types and functions to represent diagonals, codiagonals, duals, caps, cups, daggers, and other structures in wiring diagrams.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.BoxOp — Type

Box wrapping another box.

Represents unary operations on boxes in wiring diagrams.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.Junction — Type

Junction node in a wiring diagram.

Junction nodes are used to explicitly represent copies, merges, deletions, creations, caps, and cups.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.PortOp — Type

Port value wrapping another value.

Represents unary operations on ports in wiring diagrams.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.Ports — Type

A list of ports.

The objects in categories of wiring diagrams.

Catlab.Syntax.functor — Method

Apply functor in a category of wiring diagrams.

Defined by compatible mappings of ports and boxes.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.add_junctions — Method

Add junction nodes to wiring diagram.

Transforms from the implicit to the explicit representation of diagonals and codiagonals. This operation is inverse to rem_junctions.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.junction_caps — Method

Wiring diagram of nested caps made out of junction nodes.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.junction_cups — Method

Wiring diagram of nested cups made out of junction nodes.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.junction_diagram — Method

Wiring diagram with a junction node for each of the given ports.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.merge_junctions — Method

Merge adjacent junction nodes into single junctions.

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.ocompose — Method

Operadic composition of wiring diagrams.

This generic function has two different signatures, corresponding to the two standard definitions of an operad (Yau, 2018, Operads of Wiring Diagrams, Definitions 2.3 and 2.10).

This operation is a simple wrapper around substitution (substitute).

Catlab.WiringDiagrams.AlgebraicWiringDiagrams.rem_junctions — Method

Remove junction nodes from wiring diagram.

Transforms from the explicit to the implicit representation of diagonals and codiagonals. This operation is inverse to add_junctions.

Catlab.WiringDiagrams.WiringDiagramAlgorithms — Module

Algorithms operating on wiring diagrams.

Catlab.WiringDiagrams.WiringDiagramAlgorithms.crossing_minimization_by_sort — Method

Crossing minimization by sorting a univariate statistic.

The boxes in sources and/or targets are fixed and the boxes in vs are permuted. A permutation σ of the latter is returned, such that vs[σ] are the sorted box IDs. Both one-sided and two-sided crossing minimization are supported, depending on whether just one, or both, of sources and targets are given.

In this simple but popular heuristic algorithm, the boxes are permuted by sorting a univariate statistic of the positions of incoming and/or outgoing wires. Typical choices are:

mean: the sample mean, yielding the "barycenter method"

median: the sample median

In both cases, this algorithm has the property that if there is a permutation with no crossings, it will find it.

Catlab.WiringDiagrams.WiringDiagramAlgorithms.normalize_cartesian! — Method

Put a wiring diagram for a cartesian category into normal form.

This function puts a wiring diagram representing a morphism in a free cartesian category into normal form. Copies and deletions are simplified as much as possible.

Catlab.WiringDiagrams.WiringDiagramAlgorithms.normalize_copy! — Method

Normalize copies in a wiring diagram.

This function maximizes sharing of intermediate computations in a wiring diagram where copies are natural.

This algorithm is basically the same as the congruence closure algorithm on term graphs, in the special case of the empty relation R = ∅ (Baader & Nipkow, 1998, Term Rewriting and All That, Sec. 4.4). The main difference is the possibility of zero or many function outputs.

Catlab.WiringDiagrams.WiringDiagramAlgorithms.normalize_delete! — Method

Normalize deletions in a wiring diagram.

This function removes all unused intermediate computations in a wiring diagram where deletion is natural.

Catlab.WiringDiagrams.WiringDiagramAlgorithms.topological_sort — Method

Topological sort of boxes in wiring diagram.

Returns a list of box IDs, excluding the outer box's input and output IDs.

Catlab.WiringDiagrams.WiringDiagramSerialization — Module

Conventions for serialization of wiring diagrams.

Defines a consistent set of names for boxes, ports, and wires to be used when serializing wiring diagrams, as well as conventions for serializing box, port, and wire attributes.

Catlab.WiringDiagrams.GraphMLWiringDiagrams — Module

Serialize abstract wiring diagrams as GraphML.

Serialization of box, port, and wire values can be overloaded by data type (see convert_to_graphml_data and convert_from_graphml_data).

GraphML is the closest thing to a de jure and de facto standard in the space of graph data formats, supported by a variety of graph applications and libraries. We depart mildly from the GraphML spec by allowing JSON data attributes for GraphML nodes, ports, and edges.

References:

GraphML Primer: http://graphml.graphdrawing.org/primer/graphml-primer.html

GraphML DTD: http://graphml.graphdrawing.org/specification/dtd.html

Catlab.WiringDiagrams.GraphMLWiringDiagrams.generate_graphml — Method

Generate GraphML representing a wiring diagram.

Catlab.WiringDiagrams.GraphMLWiringDiagrams.parse_graphml — Method

Parse a wiring diagram from a GraphML string or XML document.

Catlab.WiringDiagrams.GraphMLWiringDiagrams.read_graphml — Method

Read a wiring diagram from a GraphML file.

Catlab.WiringDiagrams.GraphMLWiringDiagrams.write_graphml — Method

Write a wiring diagram to a file as GraphML.

Catlab.WiringDiagrams.JSONWiringDiagrams — Module

Serialize abstract wiring diagrams as JSON.

JSON data formats are convenient when programming for the web. Unfortunately, no standard for JSON graph formats has gained any kind of widespread adoption. We adopt a format compatible with that used by the KEILER project and its successor ELK (Eclipse Layout Kernel). This format is roughly feature compatible with GraphML, supporting nested graphs and ports. It also supports layout information like node position and size.

References:

KEILER's JSON graph format: https://rtsys.informatik.uni-kiel.de/confluence/display/KIELER/JSON+Graph+Format

ELK's JSON graph format: https://www.eclipse.org/elk/documentation/tooldevelopers/graphdatastructure/jsonformat.html

Catlab.WiringDiagrams.JSONWiringDiagrams.generate_json_graph — Method

Generate a JSON dict representing a wiring diagram.

Catlab.WiringDiagrams.JSONWiringDiagrams.parse_json_graph — Method

Parse a wiring diagram from a JSON string or dict.

Catlab.WiringDiagrams.JSONWiringDiagrams.read_json_graph — Method

Read a wiring diagram from a JSON file.

Catlab.WiringDiagrams.JSONWiringDiagrams.write_json_graph — Method

Write a wiring diagram to a file as JSON.